% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_model.R
\name{cache_Q_all}
\alias{cache_Q_all}
\title{place Q_matrix elements in cache, retrieving from Bioconductor Open Storage Network}
\usage{
cache_Q_all(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache}
}
\value{
list of bfcadd results
}
\description{
place Q_matrix elements in cache, retrieving from Bioconductor Open Storage Network
}
