% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cell_type_info<-}
\alias{cell_type_info<-}
\title{Generic setter for cell_type_info slot}
\usage{
cell_type_info(object) <- value
}
\arguments{
\item{object}{An object with a cell_type_info slot}

\item{value}{The new value for the cell_type_info slot}
}
\value{
The updated object
}
\description{
Generic setter for cell_type_info slot
}
\keyword{internal}
