% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cell_type_info}
\alias{cell_type_info}
\alias{cell_type_info,RctdConfig-method}
\alias{cell_type_info<-,RctdConfig-method}
\title{Generic accessor for cell_type_info slot}
\usage{
cell_type_info(object)

\S4method{cell_type_info}{RctdConfig}(object)

\S4method{cell_type_info}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with a cell_type_info slot}
}
\value{
The cell_type_info slot of the object
}
\description{
Generic accessor for cell_type_info slot
}
\keyword{internal}
