% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cell_types<-}
\alias{cell_types<-}
\title{Generic setter for cell_types slot}
\usage{
cell_types(object) <- value
}
\arguments{
\item{object}{An object with a cell_types slot}

\item{value}{The new value for the cell_types slot}
}
\value{
The updated object
}
\description{
Generic setter for cell_types slot
}
\keyword{internal}
