% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cell_types}
\alias{cell_types}
\alias{cell_types,Reference-method}
\alias{cell_types<-,Reference-method}
\title{Generic accessor for cell_types slot}
\usage{
cell_types(object)

\S4method{cell_types}{Reference}(object)

\S4method{cell_types}{Reference}(object) <- value
}
\arguments{
\item{object}{An object with a cell_types slot}
}
\value{
The cell_types slot of the object
}
\description{
Generic accessor for cell_types slot
}
\keyword{internal}
