% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{choose_sigma_c}
\alias{choose_sigma_c}
\title{Estimates sigma_c by maximum likelihood}
\usage{
choose_sigma_c(iv, resultsdir)
}
\arguments{
\item{iv}{Initial Variables: meta data obtained from the \code{\link{init_RCTD}} function}

\item{resultsdir:}{directory of results folder}
}
\value{
returns sigma, a numeric representing estimated sigma_c
(the predicted bulk cell type proportions)
}
\description{
Estimates sigma_c by stochastic gradient descent
}
