% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{config<-}
\alias{config<-}
\title{Generic setter for config slot}
\usage{
config(object) <- value
}
\arguments{
\item{object}{An object with a config slot}

\item{value}{The new value for the config slot}
}
\value{
The updated object
}
\description{
Generic setter for config slot
}
\keyword{internal}
