% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{config}
\alias{config}
\alias{config,RctdConfig-method}
\alias{config<-,RctdConfig-method}
\title{Generic accessor for config slot}
\usage{
config(object)

\S4method{config}{RctdConfig}(object)

\S4method{config}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with a config slot}
}
\value{
The config slot of the object
}
\description{
Generic accessor for config slot
}
\keyword{internal}
