% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{convert.old.RCTD}
\alias{convert.old.RCTD}
\title{Updates an old \code{\linkS4class{RCTD}} object to be compatible with the current
version of \code{spacexr}.}
\usage{
convert.old.RCTD(myRCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object (potentially from an older version.}
}
\value{
an \code{\linkS4class{RCTD}} object updated to be compatible with the current version
of \code{spacexr}.
}
\description{
Updates an old \code{\linkS4class{RCTD}} object to be compatible with the current
version of \code{spacexr}.
}
