% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coords<-}
\alias{coords<-}
\title{Generic setter for coords slot}
\usage{
coords(object) <- value
}
\arguments{
\item{object}{An object with a coords slot}

\item{value}{The new value for the coords slot}
}
\value{
The updated object
}
\description{
Generic setter for coords slot
}
\keyword{internal}
