% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coords}
\alias{coords}
\title{Get spatial coordinates}
\usage{
coords(x)
}
\arguments{
\item{x}{A SpatialRNA object}
}
\value{
A data frame containing x and y coordinates for each pixel
}
\description{
Get spatial coordinates
}
\examples{
data(rctd_simulation)

spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)

# Get coordinates for all pixels
pixel_coords <- coords(spatial_rna)
head(pixel_coords)

x_coords <- pixel_coords$x
y_coords <- pixel_coords$y

}
