% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_utils.R
\name{count_cell_types}
\alias{count_cell_types}
\title{Counts number of pixel occurrences for each cell type to be used in the CSIDE model}
\usage{
count_cell_types(
  myRCTD,
  barcodes,
  cell_types,
  cell_type_threshold = 125,
  doublet_mode = T,
  weight_threshold = NULL
)
}
\arguments{
\item{barcodes}{the barcodes, or pixel names, of the \code{\linkS4class{SpatialRNA}} object to be used when counting cell typel\.}

\item{cell_types}{the cell types used for CSIDE. If null, cell types will be chosen with aggregate occurences of
at least `cell_type_threshold`, as aggregated by \code{\link{aggregate_cell_types}}}

\item{cell_type_threshold}{(default 125) min occurence of number of cells for each cell type to be used, as aggregated by \code{\link{aggregate_cell_types}}}

\item{doublet_mode}{(default TRUE) if TRUE, uses RCTD doublet mode weights. Otherwise, uses RCTD full mode weights}

\item{weight_threshold}{(default NULL) the threshold of total normalized weights across all cell types
in \code{cell_types} per pixel to be included in the model. Default 0.99 for doublet_mode or 0.95 for full_mode.}

\item{RCTD}{an \code{\linkS4class{RCTD}} object with annotated cell types e.g. from the \code{\link{run.RCTD}} function.}
}
\value{
a named vector of number of pixel occurrences for each cell type
}
\description{
The difference with \code{\link{aggregate_cell_types}} is that this function does filter out pixels
based on total cell type weight, as occurs in the CSIDE algorithm.
}
