% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{counts<-}
\alias{counts<-}
\title{Generic setter for counts slot}
\usage{
counts(object) <- value
}
\arguments{
\item{object}{An object with a counts slot}

\item{value}{The new value for the counts slot}
}
\value{
The updated object
}
\description{
Generic setter for counts slot
}
\keyword{internal}
