% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{counts}
\alias{counts}
\alias{counts,Reference-method}
\title{Get gene expression counts}
\usage{
counts(x)

\S4method{counts}{Reference}(x)
}
\arguments{
\item{x}{A SpatialRNA or Reference object}
}
\value{
A sparse matrix of gene expression counts
}
\description{
Get gene expression counts
}
\examples{
data(rctd_simulation)

## SpatialRNA Example
spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)

spatial_rna_counts <- counts(spatial_rna)
dim(spatial_rna_counts)

# Basic operations with counts matrices
total_genes <- nrow(spatial_rna_counts)
total_pixels <- ncol(spatial_rna_counts)
gene_names <- rownames(spatial_rna_counts)

## Reference Example
reference <- Reference(
    rctd_simulation$reference_counts,
    rctd_simulation$reference_cell_types
)

reference_counts <- counts(reference)
dim(reference_counts)

}
