% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{create.RCTD.unsupervised}
\alias{create.RCTD.unsupervised}
\title{Creates an \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Reference} object and a \code{\linkS4class{SpatialRNA}} object}
\usage{
create.RCTD.unsupervised(
  spatialRNA,
  cell_type_info,
  max_cores = 4,
  gene_list = NULL,
  gene_cutoff_reg = 2e-04,
  fc_cutoff_reg = 0.75,
  UMI_min = 100,
  UMI_max = 2e+07,
  UMI_min_sigma = 300,
  class_df = NULL,
  CONFIDENCE_THRESHOLD = 10,
  DOUBLET_THRESHOLD = 25
)
}
\arguments{
\item{spatialRNA}{a \code{\linkS4class{SpatialRNA}} object to run RCTD on}

\item{cell_type_info}{a \code{cell_type_info} initialization}

\item{max_cores}{for parallel processing, the number of cores used. If set to 1, parallel processing is not used. The system will additionally be checked for
number of available cores.}

\item{gene_list}{(optional) complete list of genes to be included.}

\item{gene_cutoff_reg}{minimum normalized gene expression for genes to be included.}

\item{fc_cutoff_reg}{minimum log-fold-change (across cell types) for genes to be included.}

\item{UMI_min}{minimum UMI per pixel included in the analysis}

\item{UMI_max}{maximum UMI per pixel included in the analysis}

\item{UMI_min_sigma}{minimum UMI per pixel for the \link{choose_sigma_c} function}

\item{class_df}{(optional) if not NULL, then a dataframe mapping each cell type to a cell class, so that RCTD will report confidence on the class level.}

\item{CONFIDENCE_THRESHOLD}{(Default 10) the minimum change in likelihood (compared to other cell types) necessary to determine a cell type identity with confidence}

\item{DOUBLET_THRESHOLD}{(Default 25) the penalty weight of predicting a doublet instead of a singlet for a pixel}
}
\value{
an \code{\linkS4class{RCTD}} object, which is ready to run the \code{\link{run.RCTD}} function
}
\description{
Creates an \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Reference} object and a \code{\linkS4class{SpatialRNA}} object
}
