% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{create_cell_type_info}
\alias{create_cell_type_info}
\title{Create cell type information}
\usage{
create_cell_type_info(
  reference = NULL,
  cell_type_names = NULL,
  cell_type_profiles = NULL,
  ref_n_cells_min = 25
)
}
\arguments{
\item{reference}{\code{\linkS4class{Reference}} object or NULL if using
cell_type_profiles}

\item{cell_type_names}{character vector of cell type names to include,
optional}

\item{cell_type_profiles}{matrix of precomputed cell type expression profiles
(genes by cell type), optional}

\item{ref_n_cells_min}{numeric, minimum number of cells per cell type in the
reference (default: 25)}
}
\value{
A list containing cell type information
}
\description{
Create cell type information
}
\keyword{internal}
