% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{create_puck_and_internal_vars}
\alias{create_puck_and_internal_vars}
\title{Create puck and internal variables}
\usage{
create_puck_and_internal_vars(
  spatialRNA,
  cell_type_info,
  config,
  class_df = NULL
)
}
\arguments{
\item{spatialRNA}{\code{\linkS4class{SpatialRNA}} object}

\item{cell_type_info}{list containing cell type information}

\item{config}{list of configuration parameters}

\item{class_df}{data frame mapping cell types to classes, optional}
}
\value{
List containing the puck and internal variables
}
\description{
Create puck and internal variables
}
\keyword{internal}
