% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{create_se_doublet}
\alias{create_se_doublet}
\title{Converts the results of \code{process_beads_batch} to a SummarizedExperiment}
\usage{
create_se_doublet(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{\code{process_beads_batch} results}
}
\description{
Converts the results of \code{process_beads_batch} to a SummarizedExperiment
}
\keyword{internal}
