% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{create_se_full}
\alias{create_se_full}
\title{Converts the results of \code{decompose_batch} to a SummarizedExperiment}
\usage{
create_se_full(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{\code{decompose_batch} results}
}
\description{
Converts the results of \code{decompose_batch} to a SummarizedExperiment
}
\keyword{internal}
