% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{create_se_multi}
\alias{create_se_multi}
\title{Converts the results of \code{process_beads_multi} to a SummarizedExperiment}
\usage{
create_se_multi(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{\code{process_beads_multi} results}
}
\value{
SummarizedExperiment containing RCTD results
}
\description{
Converts the results of \code{process_beads_multi} to a SummarizedExperiment
}
\keyword{internal}
