% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{create_spe_doublet}
\alias{create_spe_doublet}
\title{Converts the results of \code{process_beads_batch} to a
\code{SpatialExperiment}}
\usage{
create_spe_doublet(RCTD, results)
}
\arguments{
\item{RCTD}{RctdConfig object}

\item{results}{\code{process_beads_batch} results}
}
\value{
\code{SpatialExperiment} containing RCTD results
}
\description{
Converts the results of \code{process_beads_batch} to a
\code{SpatialExperiment}
}
\keyword{internal}
