% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{create_spe_full}
\alias{create_spe_full}
\title{Converts the results of \code{decompose_batch} to a \code{SpatialExperiment}}
\usage{
create_spe_full(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{\code{decompose_batch} results}
}
\value{
\code{SpatialExperiment} containing RCTD results
}
\description{
Converts the results of \code{decompose_batch} to a \code{SpatialExperiment}
}
\keyword{internal}
