% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{create_spe_multi}
\alias{create_spe_multi}
\title{Converts the results of \code{process_beads_multi} to a
\code{SpatialExperiment}}
\usage{
create_spe_multi(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{\code{process_beads_multi} results}
}
\value{
\code{SpatialExperiment} containing RCTD results
}
\description{
Converts the results of \code{process_beads_multi} to a
\code{SpatialExperiment}
}
\keyword{internal}
