% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeToSeurat.R
\name{dgeToSeurat}
\alias{dgeToSeurat}
\title{Creates a scRNA-seq Seurat object}
\usage{
dgeToSeurat(refdir)
}
\arguments{
\item{refdir}{(string) the directory of the reference}
}
\value{
Returns a \code{\link{Seurat}} object containing the scRNA-seq reference, downsampled to 1000 cells
(per cell type).
}
\description{
Given a reference directory folder with 3 files: \code{dge.csv}, \code{meta_data.csv},
and \code{cell_type_dict.csv}. Saves the reference as refdir/'SCRef.RDS' and also
saves a reference downsampled to 1000 cells (per cell type) as refdir/'scRefSubsampled.RDS'
}
\section{Input file format (contained in refdir)}{

\enumerate{
\item \code{meta_data.csv} # a CSV file (with 3 columns, with headers "barcode", "cluster", and "nUMI") containing the numeric cluster assignment for each cell.
\item \code{cell_type_dict.csv} # a CSV file (with 2 columns, with headers "Cluster" and "Name") containing the mapping between numeric cluster ID and cluster name. If you want a cluster to be filtered out of the single cell reference, you can leave the cluster name blank. The cell types must not contain the character '/' or '-'.
\item \code{dge.csv} # a DGE (barcodes by gene counts) CSV file in the standard 10x format.
}
}

