% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_utils.R
\name{exvar.point.density}
\alias{exvar.point.density}
\title{Constructs an explanatory variable representing density of a set of points}
\usage{
exvar.point.density(myRCTD, barcodes, points, radius = 50)
}
\arguments{
\item{myRCTD}{an \code{\linkS4class{RctdConfig}} object adapted from the
output of the \code{\link{runRctd}} function.}

\item{barcodes}{the barcodes, or pixel names, of the
\code{\linkS4class{SpatialRNA}} for which to evaluate the explanatory
variable. These would be the pixels used in the C-SIDE model.}

\item{points}{a N by 2 matrix containing the locations of the points to be
used for computing density. The first column should be the x coordinates
while the second column should be the y coordinate.}

\item{radius}{(default 50) the radius of the exponential filter.
Approximately, the distance considered to be a relevant interaction.}
}
\value{
explanatory.variable a named numeric vector representing the
  explanatory variable used for explaining differential expression in CSIDE.
  Names of the variable are the \code{\linkS4class{SpatialRNA}} pixel names,
  and values  are standardized between 0 and 1. This variable represents
  density of the given point set.
}
\description{
This explanatory variable can be used with CSIDE to detect DE in the
proximity of these points. Density is computing using an
exponentially-decaying filter.
}
