% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{fitBulk}
\alias{fitBulk}
\title{Performs Platform Effect Normalization:}
\usage{
fitBulk(iv)
}
\arguments{
\item{iv}{Initial Variables: meta data obtained from the \code{\link{init_RCTD}} function}
}
\value{
Returns a list of cell_type_info_renorm (the normalized cell type profiles) and proportions
(the predicted bulk cell type proportions)
}
\description{
Estimates bulk cell type composition and uses this
to estimate platform effects and normalize cell type proportions
}
\section{Additional Output Files}{

Saves proportions in results/Bulk/weights.RDS.
Saves cell_type_info_renorm in MetaData/cell_type_info_renorm.RDS
Additionally, creates trace files of sigma fitting in results/sigmaBulk
}

