% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{gather_results}
\alias{gather_results}
\title{Collects RCTD results}
\usage{
gather_results(puck, iv)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{iv}{Initial Variables: meta data obtained from the \code{\link{init_RCTD}} function}
}
\value{
Returns a list of results_df (a dataframe of RCTD results in doublet mode),
weights (a dataframe of RCTD predicted weights in full mode), and weights_doublet (a
dataframe of predicted weights in doublet mode, with cell type information in results_df)
}
\description{
Assumes that RCTD has saved results in /SplitPuckResults/results
for each fold index. Collects and organizes these results.
}
\section{results_df output format}{

Contains one column per pixel (barcodes as rownames). Important columns
are spot_class, a factor variable representing RCTD's classification in
doublet mode: "singlet" (1 cell type on pixel), "doublet_certain" (2 cell types
on pixel), "doublet_uncertain" (2 cell types on pixel, but only confident of 1),
"reject" (no prediction given for pixel). Next, the first_type column gives the
first cell type predicted on the bead (for all spot_class conditions except "reject").
The second_type column gives the second cell type predicted on the bead for doublet
spot_class conditions (not a confident prediction for "doublet_uncertain").
}

