% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_class_df}
\alias{get_class_df}
\title{Returns a dataframe mapping cell types to classes}
\usage{
get_class_df(cell_type_names, use_classes = FALSE)
}
\arguments{
\item{cell_type_names}{List of cell type names}

\item{use_classes}{Whether to use pre-defined classes}
}
\value{
Dataframe mapping cell types to classes. If \code{use_classes} is
  \code{FALSE}, then each cell type will correspond to its own class.
  Otherwise, similar cell types will be grouped together according to a
  pre-defined map.
}
\description{
Returns a dataframe mapping cell types to classes
}
\keyword{internal}
