% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{get_decomposed_data}
\alias{get_decomposed_data}
\title{Decomposes SpatialRNA data into individual cells}
\usage{
get_decomposed_data(
  results_df,
  gene_list,
  puck,
  weights_doublet,
  cell_type_info
)
}
\arguments{
\item{results_df}{a dataframe of RCTD results}

\item{gene_list}{a list of genes to be used for the decomposition}

\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{weights_doublet}{a dataframe of predicted weights in doublet mode}

\item{cell_type_info}{cell type information and profiles of each cell, calculated from the scRNA-seq
reference (see \code{\link{get_cell_type_info}})}
}
\value{
An object of type \linkS4class{SpatialRNA} representing the decomposed cells
}
\description{
Warning: in the current RCTD version, this function is deprecated, and is no longer supported.
For differential expression tasks, we instead recommend the RCTDE method.
}
\details{
Applied to the output of \code{\link{gather_results}}.
Singlet pixels are left unchanged, and doublet_certain conditions are
decomposed into single cells.
}
