% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_utils.R
\name{get_doublet_weights}
\alias{get_doublet_weights}
\title{Converts RCTD doublet mode results to a weight matrix (across all cell types)}
\usage{
get_doublet_weights(myRCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object with annotated cell types from the \code{\link{run.RCTD}} function.}
}
\value{
a weights matrix of cell type proportions for each pixel and each cell type.
}
\description{
RCTD must have been run in doublet mode
}
