% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE_utils.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Generates marker genes from C-SIDE means on a fitted
\code{\linkS4class{RctdConfig}} object}
\usage{
get_marker_genes(
  RCTD,
  cell_types = NULL,
  fdr = 0.05,
  fc_thresh = 1.25,
  expr_thresh = 0.00015,
  max_sigma = 1
)
}
\arguments{
\item{RCTD}{a \code{\linkS4class{RctdConfig}} object}

\item{cell_types}{the cell types used to find marker genes. If null, all cell
types in \code{RCTD@internal_vars_de$cell_types} will be chosen.}

\item{fdr}{false discovery rate for marker genes.}

\item{fc_thresh}{minimum \code{log_e} fold change required for a gene to be
considered a marker gene.}

\item{expr_thresh}{minimum expression threshold, as normalized expression
(proportion out of 1, or counts per 1).}

\item{max_sigma}{maximum standard error of gene expression mean.}
}
\value{
a dataframe of marker gene names, expressions, and p-values for each
  cell type
}
\description{
Generates marker genes from C-SIDE means on a fitted
\code{\linkS4class{RctdConfig}} object
}
