% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_info.R
\name{get_norm_ref}
\alias{get_norm_ref}
\title{Normalizes cell type profiles to a target dataset}
\usage{
get_norm_ref(puck, cell_type_means, gene_list, proportions)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}, the target dataset}

\item{cell_type_means}{a data_frame (genes by cell types) for mean normalized
expression (see \code{\link{get_cell_type_info}})}

\item{gene_list}{a list of genes to be used for the normalization}

\item{proportions}{a named list (for each cell type) of proportion of the
cell type on the bulk dataset (not constrained to sum to 1)}
}
\value{
Returns \code{cell_type_means}, a data_frame (genes by cell types)
  for mean normalized cell type expression profiles in which platform effects
  have been removed to match the \linkS4class{SpatialRNA} data.
}
\description{
Renormalizes \code{cell_type_means} to have average the same as the puck. The
average for each gene is weighted by cell type proportions given by
\code{proportions}.
}
