% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_model.R
\name{get_Q_mat}
\alias{get_Q_mat}
\title{Calculates Poisson-lognormal Likelihood}
\usage{
get_Q_mat(iv, sigma, big_params = T)
}
\arguments{
\item{iv}{Initial Variables: meta data obtained from the \code{\link{init_RCTD}} function}

\item{sigma}{a numeric representing estimated sigma_c from \code{\link{choose_sigma_c}}}
}
\value{
Returns \code{Q_mat}, a matrix containing pre-computed values for the log likelihood.
}
\description{
For a given value of sigma, calculates P(y|x) under the Poisson-Lognormal model
for each possible value of y and x.
}
