% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{import_weights}
\alias{import_weights}
\title{Assigns a cell type `weights` matrix to an \code{\linkS4class{RCTD}} object}
\usage{
import_weights(myRCTD, weights)
}
\arguments{
\item{myRCTD}{a \code{\linkS4class{RCTD}} object to be assigned weights.}

\item{weights}{a matrix of weights (pixels by cell types). weights must be normalized
to have rows sum to 1. Furthermore, rownames and colnames must be assigned as
pixel names and cell types respectively.}
}
\value{
the \code{\linkS4class{RCTD}} object with weights assigned.
}
\description{
Assigns a cell type `weights` matrix to an \code{\linkS4class{RCTD}} object
}
