% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_RCTD}
\alias{init_RCTD}
\title{Initializes RCTD, returns meta-data}
\usage{
init_RCTD(
  gene_list_reg = T,
  get_proportions = F,
  test_reference = NULL,
  puck_file = NULL,
  MIN_OBS = 1,
  load_info_renorm = F,
  load_info = F
)
}
\arguments{
\item{gene_list_reg}{logical, if T uses the 'reg' cutoffs for gene list, rather
than the bulk cutoffs.}

\item{get_proportions}{logical, whether to load proportions from 'results/Bulk/weights.RDS'}

\item{load_info_renorm}{logical, if true, reads \code{cell_type_info} from \code{MetaData/cell_type_info_renorm.RDS}}
}
\value{
A named list of meta data, including
\itemize{
  \item refdir, the directory of the scRNA-seq reference
  \item SpatialRNAdir, the directory of the SpatialRNA data
  \item bulkdir, the directory of results for the bulk analysis
  \item reference, a \code{\link{Seurat}} object of the scRNA-seq reference
  \item proportions, a a named list (for each cell type) of proportion of the cell type on the bulk dataset
  \item gene_list, a list of genes to be used for the normalization
  \item puck, an object of type \linkS4class{SpatialRNA}
  \item cell_type_info, cell type information and profiles of each cell, calculated from the scRNA-seq
reference (see \code{\link{get_cell_type_info}})
  \item config, additional configuration parameters
  \item n_puck_folds, the number of folds to split the puck for RCTD
  \item puckrds, the filename of the SpatialRNA object
}
}
\description{
Returns a named list of data used for running RCTD. Reads in config file
\code{conf/dataset.yml}, and then reads in the conf/config_mode.yml file
using the config_mode from the first config file. These config files must be configured
to point RCTD to the correct datasets, and to provide parameters for RCTD.
Will search for scRNA-seq reference in 'reffolder/reffile'.
Will search for SpatialRNA data in 'SpatialRNAfolder/puckrds'
}
