% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{internal_vars<-}
\alias{internal_vars<-}
\title{Generic setter for internal_vars slot}
\usage{
internal_vars(object) <- value
}
\arguments{
\item{object}{An object with an internal_vars slot}

\item{value}{The new value for the internal_vars slot}
}
\value{
The updated object
}
\description{
Generic setter for internal_vars slot
}
\keyword{internal}
