% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{internal_vars}
\alias{internal_vars}
\alias{internal_vars,RctdConfig-method}
\alias{internal_vars<-,RctdConfig-method}
\title{Generic accessor for internal_vars slot}
\usage{
internal_vars(object)

\S4method{internal_vars}{RctdConfig}(object)

\S4method{internal_vars}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with an internal_vars slot}
}
\value{
The internal_vars slot of the object
}
\description{
Generic accessor for internal_vars slot
}
\keyword{internal}
