% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{internal_vars_de<-}
\alias{internal_vars_de<-}
\title{Generic setter for internal_vars_de slot}
\usage{
internal_vars_de(object) <- value
}
\arguments{
\item{object}{An object with an internal_vars_de slot}

\item{value}{The new value for the internal_vars_de slot}
}
\value{
The updated object
}
\description{
Generic setter for internal_vars_de slot
}
\keyword{internal}
