% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{internal_vars_de}
\alias{internal_vars_de}
\alias{internal_vars_de,RctdConfig-method}
\alias{internal_vars_de<-,RctdConfig-method}
\title{Generic accessor for internal_vars_de slot}
\usage{
internal_vars_de(object)

\S4method{internal_vars_de}{RctdConfig}(object)

\S4method{internal_vars_de}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with an internal_vars_de slot}
}
\value{
The internal_vars_de slot of the object
}
\description{
Generic accessor for internal_vars_de slot
}
\keyword{internal}
