% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE.R
\name{iterOptim}
\alias{iterOptim}
\title{Runs the unsupervised iterative optimization algorithm}
\usage{
iterOptim(
  RCTD,
  cell_types,
  doublet_mode = "doublet",
  n_iter = 50,
  MIN_CHANGE = 0.001,
  initialSol = NULL,
  return_list = FALSE
)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the
\code{\link{choose_sigma_c}} function.}

\item{cell_types}{the cell types used for CSIDE.}

\item{doublet_mode}{\code{character string}, either "doublet", "subtype", or
"full" on which mode to run iterOptim. Please see above description.}

\item{n_iter}{maximum number of optimization iterations}

\item{MIN_CHANGE}{minimum change required to terminate optimization}
}
\value{
an \code{\linkS4class{RCTD}} object containing the results of the
  unsupervised algorithm.
}
\description{
If in doublet mode, fits at most two cell types per pixel. It classifies each
pixel as 'singlet' or 'doublet' and searches for the cell types on the pixel.
If in full mode, can fit any number of cell types on each pixel. If in
subtype mode, fits only pixels belonging to given cell type(s).
}
