% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLAMDE_plots.R
\name{make_all_de_plots}
\alias{make_all_de_plots}
\title{Makes all GLAMDE plots on RCTD object, after running GLAMDE}
\usage{
make_all_de_plots(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running GLAMDE}

\item{datadir}{output directory}
}
\description{
Makes all GLAMDE plots on RCTD object, after running GLAMDE
}
