% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDE_plots.R
\name{make_de_plots_quant}
\alias{make_de_plots_quant}
\title{Makes quantitative RCTDE plots on RCTD object, after running RCTDE}
\usage{
make_de_plots_quant(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running RCTDE}

\item{datadir}{output directory}
}
\description{
Makes quantitative RCTDE plots on RCTD object, after running RCTDE
}
