% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLAMDE_plots.R
\name{make_de_plots_regions}
\alias{make_de_plots_regions}
\title{Makes spatial gene GLAMDE plots (colored by discrete regions) on RCTD object, after running GLAMDE}
\usage{
make_de_plots_regions(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running GLAMDE}

\item{datadir}{output directory}
}
\description{
Makes spatial gene GLAMDE plots (colored by discrete regions) on RCTD object, after running GLAMDE
}
