% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{make_de_plots_replicates}
\alias{make_de_plots_replicates}
\title{Makes spatial gene CSIDE plots (colored by two discrete regions) on RCTD replicates object, after running CSIDE}
\usage{
make_de_plots_replicates(RCTD.replicates, datadir)
}
\arguments{
\item{RCTD.replicates}{a \code{\linkS4class{RCTD.replicates}} object after performing population-level DE inference.}

\item{datadir}{output directory}
}
\description{
Runs on genes that were identified as significant at the population level
}
\details{
These plots are colored by two discrete regions based on high or low explanatory variable values.
Bold points represent expressed, whereas unbold points represent pixels not expressing the gene.
}
