% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDreplicates.R
\name{merge.RCTD.objects}
\alias{merge.RCTD.objects}
\title{Creates an \code{\linkS4class{RCTD.replicates}} object across multiple \code{\linkS4class{RCTD}} objects}
\usage{
\method{merge}{RCTD.objects}(RCTD.reps, replicate_names, group_ids = NULL)
}
\arguments{
\item{RCTD.reps}{a list of multiple \code{\linkS4class{RCTD}} objects to merge into one \code{\linkS4class{RCTD.replicates}} object.}

\item{replicate_names}{a \code{character} vector of names for each replicate provided in \code{RCTD.reps}}

\item{group_ids}{(default constant across replicates) a named integer vector (length number of replicates) containing the group id for each replicate.
Names represent the replicate names, and replicates of the same group will be expected to be more similar than
replicates across groups}
}
\value{
an \code{\linkS4class{RCTD.replicates}} object, containing each \code{\linkS4class{RCTD}} object in \code{RCTD.reps}
}
\description{
Creates an \code{\linkS4class{RCTD.replicates}} object across multiple \code{\linkS4class{RCTD}} objects
}
