% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{normalize_weights}
\alias{normalize_weights}
\title{Normalizes the `weights` matrix from the RCTD results object}
\usage{
normalize_weights(weights)
}
\arguments{
\item{weights}{a matrix of weights to be normalized}
}
\value{
norm.weights a normalized matrix of weights where rows sum to one.
}
\description{
Normalizes the `weights` matrix from the RCTD results object
}
