% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{nUMI<-}
\alias{nUMI<-}
\title{Generic setter for nUMI slot}
\usage{
nUMI(object) <- value
}
\arguments{
\item{object}{An object with a nUMI slot}

\item{value}{The new value for the nUMI slot}
}
\value{
The updated object
}
\description{
Generic setter for nUMI slot
}
\keyword{internal}
