% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{nUMI}
\alias{nUMI}
\alias{nUMI,Reference-method}
\title{Get total UMI counts}
\usage{
nUMI(x)

\S4method{nUMI}{Reference}(x)
}
\arguments{
\item{x}{A SpatialRNA or Reference object}
}
\value{
A numeric vector of total UMI counts per pixel/cell
}
\description{
Get total UMI counts
}
\examples{
data(rctd_simulation)

## SpatialRNA Example
spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)

# Get UMI counts per pixel
pixel_umis <- nUMI(spatial_rna)
head(pixel_umis)
summary(pixel_umis)  # Basic statistics

## Reference Example
reference <- Reference(
    rctd_simulation$reference_counts,
    rctd_simulation$reference_cell_types
)

# Get UMI counts per cell in reference
cell_umis <- nUMI(reference)
head(cell_umis)
summary(cell_umis)  # Basic statistics

}
