% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{originalSpatialRNA<-}
\alias{originalSpatialRNA<-}
\title{Generic setter for originalSpatialRNA slot}
\usage{
originalSpatialRNA(object) <- value
}
\arguments{
\item{object}{An object with an originalSpatialRNA slot}

\item{value}{The new value for the originalSpatialRNA slot}
}
\value{
The updated object
}
\description{
Generic setter for originalSpatialRNA slot
}
\keyword{internal}
