% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{originalSpatialRNA}
\alias{originalSpatialRNA}
\alias{originalSpatialRNA,RctdConfig-method}
\alias{originalSpatialRNA<-,RctdConfig-method}
\title{Generic accessor for originalSpatialRNA slot}
\usage{
originalSpatialRNA(object)

\S4method{originalSpatialRNA}{RctdConfig}(object)

\S4method{originalSpatialRNA}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with an originalSpatialRNA slot}
}
\value{
The originalSpatialRNA slot of the object
}
\description{
Generic accessor for originalSpatialRNA slot
}
\keyword{internal}
