% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_cell_types}
\alias{plot_all_cell_types}
\title{Plots all cell types in space}
\usage{
plot_all_cell_types(results_df, coords, cell_type_names, resultsdir)
}
\arguments{
\item{results_df}{a dataframe of RCTD results (see \code{\link{gather_results}})}

\item{coords}{a dataframe of coordinates of each pixel}

\item{cell_type_names}{list of cell type names}

\item{resultsdir}{output directory}
}
\value{
returns \code{\link{ggplot2}} object
}
\description{
Plots the first cell type in doublet mode. Saves as 'all_cell_types.pdf'
}
