% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_weights}
\alias{plot_all_weights}
\title{Plot pie charts of cell type proportions across pixels}
\usage{
plot_all_weights(
  rctd_se,
  weights_assay = "weights",
  cell_type_colors = grDevices::rainbow(ncol(assay(rctd_se))),
  r = 0.4,
  lwd = 1,
  plot_title = NA
)
}
\arguments{
\item{rctd_se}{SummarizedExperiment containing RCTD results}

\item{weights_assay}{character(1) name of the weights assay (default:
"weights")}

\item{cell_type_colors}{vector of colors for the different cell types
(default: rainbow)}

\item{r}{numeric(1) radius of the pie charts (default: 0.4)}

\item{lwd}{numeric(1) line width of the pie chart borders (default: 1)}

\item{plot_title}{character(1) plot title. (default: NA)}
}
\value{
ggplot object showing cell type proportions at each pixel using pie
  charts
}
\description{
Generates a visualization where each pixel is represented by a pie chart
showing the proportions of different cell types at that location.
}
