% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_class}
\alias{plot_class}
\title{Plots a factor variable in space on the puck}
\usage{
plot_class(puck, barcodes_cur, my_class, counter_barcodes = NULL, title = NULL)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{barcodes_cur}{a list of barcodes to include in the plot}

\item{my_class}{a named (by barcode) factor list for the coloring}

\item{counter_barcodes}{(optional) list of barcodes to highlight with low
opacity}

\item{title}{the title of the plot}
}
\value{
Returns a \link{ggplot} object
}
\description{
Colors points based on class
}
\keyword{internal}
