% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cond_occur}
\alias{plot_cond_occur}
\title{Barplot of the confident counts for each cell type}
\usage{
plot_cond_occur(cell_type_info, resultsdir, weights)
}
\arguments{
\item{cell_type_info}{cell type information and profiles (see \code{\link{get_cell_type_info}})}

\item{resultsdir}{output directory}

\item{weights}{a dataframe of RCTD output weights (see \code{\link{gather_results}})}
}
\value{
returns \code{\link{ggplot2}} object
}
\description{
Plots the number of confident labels in 'full_mode'. Saves as 'cell_type_occur.pdf'
}
