% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_doublets}
\alias{plot_doublets}
\title{Plots all doublets in space}
\usage{
plot_doublets(puck, doublets, resultsdir, cell_type_names)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{doublets}{a dataframe of RCTD results restricted to doublets}

\item{resultsdir}{output directory}

\item{cell_type_names}{list of cell type names}
}
\value{
returns \code{\link{ggplot2}} object
}
\description{
saves as 'all_doublets.pdf'
}
