% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_doublets_type}
\alias{plot_doublets_type}
\title{Plots doublets of each cell type individually}
\usage{
plot_doublets_type(puck, doublets_base, resultsdir, cell_type_info)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{doublets_base}{a dataframe of RCTD results restricted to doublets}

\item{resultsdir}{output directory}

\item{cell_type_info}{cell type information and profiles (see \code{\link{get_cell_type_info}})}
}
\description{
Plots the first cell type in doublet mode. Saves as 'all_doublets_type.pdf'
}
