% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_gene_two_regions}
\alias{plot_gene_two_regions}
\title{Makes a spatial plot of differential expression for a particular gene}
\usage{
plot_gene_two_regions(
  puck,
  all_barc,
  my_beta,
  X2,
  res_genes,
  gene,
  cell_type,
  min_UMI = 200
)
}
\arguments{
\item{puck}{RCTD \code{SpatialRNA} object}

\item{gene}{gene to be plotted}
}
\description{
Makes a spatial plot of differential expression for a particular gene
}
