% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_occur_unthreshold}
\alias{plot_occur_unthreshold}
\title{Barplot of the counts for each cell type}
\usage{
plot_occur_unthreshold(cell_type_info, resultsdir, weights)
}
\arguments{
\item{cell_type_info}{cell type information and profiles (see \code{\link{get_cell_type_info}})}

\item{resultsdir}{output directory}

\item{weights}{a dataframe of RCTD output weights (see \code{\link{gather_results}})}
}
\description{
Plots the number of (including unconfident) labels in 'full_mode'. Saves as 'cell_type_occur_unthreshold.pdf'
}
