% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_puck_continuous}
\alias{plot_puck_continuous}
\title{Plots a continuous value over locations on the puck}
\usage{
plot_puck_continuous(
  puck,
  barcodes,
  plot_val,
  ylimit = c(0, 1),
  title = NULL,
  counter_barcodes = NULL,
  label = F,
  my_pal = NULL,
  xlim = NULL,
  ylim = NULL,
  size = 0.15,
  alpha = 1,
  small_point = F
)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{barcodes}{a list of barcodes to include in the plot}

\item{plot_val}{a named (by barcode) list of values to plot}

\item{ylimit}{minimum and maximum values for the range of plot as a numeric list}

\item{xlim}{(optional) minimum and maximum value for x coordinate as a numeric list}

\item{ylim}{(optional) minimum and maximum value for y coordinate as a numeric list}

\item{size}{numeric size of points}

\item{cell_type_info}{cell type information and profiles (see \code{\link{get_cell_type_info}})}
}
\value{
Returns a \link{ggplot} object
}
\description{
Colors points based on value of the function
}
