% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_puck_wrapper}
\alias{plot_puck_wrapper}
\title{Plots a continuous value over filtered locations on the puck}
\usage{
plot_puck_wrapper(
  puck,
  plot_val,
  cell_type = NULL,
  minUMI = 0,
  maxUMI = 2e+05,
  min_val = NULL,
  max_val = NULL,
  title = NULL,
  my_cond = NULL
)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{plot_val}{a named (by barcode) list of values to plot}

\item{cell_type}{string specifying cell type to plot. if NULL, then all cell types are plotted}

\item{minUMI}{numeric, minimum value for total UMIs to filter pixels}

\item{maxUMI}{numeric, maximum value for total UMIs to filter pixels}

\item{min_val}{numeric, minimum value for the range of plot as a numeric list}

\item{max_val}{numeric, maximum value for the range of plot as a numeric list}

\item{cell_type_info}{cell type information and profiles (see \code{\link{get_cell_type_info}})}
}
\value{
Returns a \link{ggplot} object
}
\description{
Colors points based on value of the function, filtered for e.g. UMI and cell type
}
