% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weights_doublet}
\alias{plot_weights_doublet}
\title{Spatially plot the weights for each cell type in doublet_mode}
\usage{
plot_weights_doublet(
  cell_type_names,
  puck,
  resultsdir,
  weights_doublet,
  results_df
)
}
\arguments{
\item{cell_type_names}{list of cell type names}

\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{resultsdir}{output directory}

\item{weights_doublet}{a dataframe of RCTD output weights for doublets (see \code{\link{gather_results}})}

\item{results_df}{dataframe of RCTD results (see \code{\link{gather_results}})}
}
\description{
Plots the weights for each cell type as in doublet_mode. Saves as 'cell_type_weights_doublet.pdf'
}
