% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weights_unthreshold}
\alias{plot_weights_unthreshold}
\title{Spatially plot all weights for each cell type}
\usage{
plot_weights_unthreshold(cell_type_names, puck, resultsdir, weights)
}
\arguments{
\item{cell_type_names}{list of cell type names}

\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{resultsdir}{output directory}

\item{weights}{a dataframe of RCTD output weights (see \code{\link{gather_results}})}
}
\description{
Plots all weights for each cell type as in full_mode. Saves as 'cell_type_weights_unthreshold.pdf'
}
