% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{process_cell_type_info}
\alias{process_cell_type_info}
\title{Process cell type information from a Reference object}
\usage{
process_cell_type_info(reference, cell_type_names, CELL_MIN = 25)
}
\arguments{
\item{reference}{\code{\linkS4class{Reference}} object}

\item{cell_type_names}{character vector of cell type names to include}

\item{CELL_MIN}{numeric, minimum number of cells per cell type in the
reference (default: 25)}
}
\value{
List containing cell type information
}
\description{
Process cell type information from a Reference object
}
\keyword{internal}
