% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRCTD.R
\name{process_data}
\alias{process_data}
\title{Runs RCTD in full mode on \code{puck}}
\usage{
process_data(
  puck,
  gene_list,
  cell_type_info,
  proportions = NULL,
  trust_model = FALSE,
  constrain = T,
  OLS = F
)
}
\arguments{
\item{puck}{an object of type \linkS4class{SpatialRNA}, the target dataset}

\item{gene_list}{a list of genes to be used for RCTD}

\item{cell_type_info}{cell type information and profiles of each cell, calculated from the scRNA-seq
reference (see \code{\link{get_cell_type_info}})}

\item{proportions}{(optional) If given, a named list (for each cell type) of proportion of the cell type on the bulk dataset
(not constrained to sum to 1)}

\item{constrain}{logical whether to constrain the weights to sum to one on each pixel}
}
\value{
Returns \code{test_results}, a list of three items:
(1) \code{conf_mat} a confusion matrix (not relevant) (2) \code{weights}
a dataframe of predicted weights (3) a named list of predicted cell types
}
\description{
Renormalizes \code{cell_type_means} to have average the same as the puck
if \code{proportions} is given. Then, computes cell type proportions for each pixel
in \code{puck}.
}
