% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_model.R
\name{Q_all_from_cache}
\alias{Q_all_from_cache}
\title{retrieve Q_mats from cache, after they have been installed using cache_Q_all}
\usage{
Q_all_from_cache(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache}
}
\value{
list of matrices
}
\description{
retrieve Q_mats from cache, after they have been installed using cache_Q_all
}
