% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RCTD-class}
\alias{RCTD-class}
\alias{show,RCTD-method}
\title{RCTD algorithm configuration}
\usage{
\S4method{show}{RCTD}(object)
}
\arguments{
\item{object}{RCTD configuration object}
}
\description{
An RCTD configuration created via the \code{\link{createRctd}} function.
Users can run RCTD by passing this object to the \code{\link{runRctd}}
function.
}
\section{Slots}{

\describe{
\item{\code{spatialRNA}}{a \code{\linkS4class{SpatialRNA}} object containing the
processed spatial transcriptomics data for analysis}

\item{\code{originalSpatialRNA}}{a \code{\linkS4class{SpatialRNA}} object containing
the complete, unfiltered spatial dataset}

\item{\code{reference}}{a \code{\linkS4class{Reference}} object containing the
annotated reference data}

\item{\code{config}}{a list of configuration options for the RCTD algorithm, set via
\code{\link{createRctd}}}

\item{\code{cell_type_info}}{a named list containing cell type expression profiles
with two components: \code{info} (raw profiles from reference data) and
\code{renorm} (profiles normalized to match the spatial data)}

\item{\code{internal_vars}}{a list of internal variables used during the RCTD
analysis}
}}

\keyword{internal}
