% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RCTD.replicates-class}
\alias{RCTD.replicates-class}
\title{An S4 class used to store multiple replicates as \code{\linkS4class{SpatialRNA}} objects.}
\description{
By storing multiple \code{\linkS4class{SpatialRNA}} replicates in this one object, it is convenient
to run RCTD and CSIDE across all replicates. Finally, multiple replicates can be combined with population-level
differential expression inference using the \code{\link{CSIDE.population.inference}} function
}
\details{
Created using the \code{\link{create.RCTD.replicates}} or \code{\link{merge.RCTD.objects}} functions.
One can run RCTD using the \code{\link{run.RCTD.replicates}} function, and one can run CSIDE using the
\code{\link{run.CSIDE.replicates}} function.
}
\section{Slots}{

\describe{
\item{\code{RCTD.reps}}{a list of \code{\linkS4class{RCTD}} objects, one for each replicate}

\item{\code{population_de_results}}{A list, indexed by cell type, of dataframes summarizing population-level
differential expression for each genes. Relevant columns include: tau, variance across replicates;
log_fc_est, the estimated differential expresison; sd_est, the standard error of estimated DE}

\item{\code{population_sig_gene_list}}{A list, indexed by cell type, of vectors of significant genes}

\item{\code{population_sig_gene_df}}{A list, indexed by cell type, of dataframe summarizing population-level
differential expression for each signifcant gene, similar to \code{population_de_results}. Additionally,
contains p (representing p-values) and q_val (representing q-values).}

\item{\code{groups_ids}}{a named integer vector (length number of replicates) containing the group id for each replicate.
Names represent the replicate names, and replicates of the same group will be expected to be more similar than
replicates across groups}
}}

