% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RctdConfig-class}
\alias{RctdConfig-class}
\alias{show,RctdConfig-method}
\title{RCTD algorithm configuration}
\usage{
\S4method{show}{RctdConfig}(object)
}
\arguments{
\item{object}{RCTD configuration object}
}
\description{
RCTD algorithm configuration
}
\section{Slots}{

\describe{
\item{\code{spatialRNA}}{a \code{\linkS4class{SpatialRNA}} object containing the
processed spatial transcriptomics data for analysis}

\item{\code{config}}{a list of configuration options for the RCTD algorithm, set via
\code{\link{createRctd}}}

\item{\code{cell_type_info}}{a named list containing cell type expression profiles
with two components: \code{info} (raw profiles from reference data) and
\code{renorm} (profiles normalized to match the spatial data)}

\item{\code{internal_vars}}{a list of internal variables used during the RCTD
analysis}

\item{\code{results}}{a \code{\link[SpatialExperiment]{SpatialExperiment}} object
containing the results of RCTD (optional, used by CSIDE)}

\item{\code{internal_vars_de}}{a list of internal variables used during the CSIDE
analysis (optional, used by CSIDE)}

\item{\code{de_results}}{a list containing the results of CSIDE (optional, used by
CSIDE)}
}}

\keyword{internal}
