% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDreplicates.R
\name{RCTDE.population.inference}
\alias{RCTDE.population.inference}
\title{Runs population-level differential expression inference for a \code{\linkS4class{RCTD.replicates}} object}
\usage{
RCTDE.population.inference(RCTD.replicates, use.groups = FALSE)
}
\arguments{
\item{RCTD.replicates}{a \code{\linkS4class{RCTD.replicates}} object for which to perform population-level DE inference.}

\item{use.groups}{(default FALSE) if TRUE, treats the replicates as having multiple groups (e.g. samples) according to the \code{group_ids} slot}
}
\value{
an \code{\linkS4class{RCTD.replicates}} object containing the results of the RCTDE population-level algorithm. See \code{\linkS4class{RCTD.replicates}}
for documentation on the \code{population_de_results}, \code{population_sig_gene_list}, and \code{population_sig_gene_df} objects.
}
\description{
First, RCTDE must have been run on all replicates using e.g. the \code{\link{run.RCTDE.replicates}} function.
}
