% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRNA.R
\name{read.VisiumSpatialRNA}
\alias{read.VisiumSpatialRNA}
\title{Creates a SpatialRNA object from a 10x Genomics Visium `outs` directory}
\usage{
read.VisiumSpatialRNA(datadir)
}
\arguments{
\item{datadir}{(string) full path to the 10x Genomics Visium `outs` directory}
}
\value{
Returns a \code{\linkS4class{SpatialRNA}} object containing the coordinates and counts
from the input files
}
\description{
Given a SpatialRNA directory 10x Genomics Visium `outs` directory and returns a SpatialRNA object.
}
