% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{reference<-}
\alias{reference<-}
\title{Generic setter for reference slot}
\usage{
reference(object) <- value
}
\arguments{
\item{object}{An object with a reference slot}

\item{value}{The new value for the reference slot}
}
\value{
The updated object
}
\description{
Generic setter for reference slot
}
\keyword{internal}
