% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{reference}
\alias{reference}
\alias{reference,RctdConfig-method}
\alias{reference<-,RctdConfig-method}
\title{Generic accessor for reference slot}
\usage{
reference(object)

\S4method{reference}{RctdConfig}(object)

\S4method{reference}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with a reference slot}
}
\value{
The reference slot of the object
}
\description{
Generic accessor for reference slot
}
\keyword{internal}
