% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRNA.R
\name{restrict_puck}
\alias{restrict_puck}
\title{Restricts a SpatialRNA object to a subset of pixels}
\usage{
restrict_puck(puck, barcodes)
}
\arguments{
\item{puck}{a \code{\linkS4class{SpatialRNA}} object}

\item{barcodes}{a list of barcode names, a subset of \code{rownames(puck@coords)}}
}
\value{
Returns a \code{\linkS4class{SpatialRNA}} object subsampled to the barcodes
}
\description{
Given a \code{\linkS4class{SpatialRNA}} object and a list of barcodes (pixels), will return a
\code{\linkS4class{SpatialRNA}} object restricted to the barcodes.
}
