% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{run.de.single}
\alias{run.de.single}
\title{Runs RCTDE on a \code{\linkS4class{RCTD}} object (for a single predictive variable)}
\usage{
run.de.single(
  myRCTD,
  explanatory.variable,
  datadir = "./",
  cell_types = NULL,
  cell_type_threshold = 125,
  gene_threshold = 5e-05,
  doublet_mode = T,
  test_mode = "direct",
  delta = 0,
  thresh_val = NULL,
  sigma_gene = T,
  PRECISION.THRESHOLD = 0.01,
  cell_types_present = NULL,
  fdr = 0.01
)
}
\arguments{
\item{explanatory.variable}{the predictive variable used for detecting differential expression in RCTDE. Names of the variable
are the `SpatialRNA` pixel names, and values should be standardized between 0 and 1.}

\item{cell_types}{the cell types used for RCTDE. If null, cell types will be chosen with aggregate occurences of
at least `cell_type_threshold`, as aggregated by \code{\link{choose_cell_types}}}

\item{cell_type_threshold}{(default 125) min occurence of number of cells for each cell type to be used, as aggregated by \code{\link{choose_cell_types}}}

\item{gene_threshold}{(default 5e-5) minimum average expression required for selecting genes}

\item{doublet_mode}{(default TRUE) if TRUE, used RCTD doublet mode weights. Otherwise, uses RCTD full mode weights}

\item{test_mode}{(default 'direct') if 'direct', tests for DE along a single parameter. If 'multi', then tests for differences
across multiple parameters}

\item{delta}{(default 0) non-centrality parameter for differential expression tests}

\item{sigma_gene}{(default TRUE) if TRUE, fits gene specific overdispersion parameter}

\item{PRECISION.THRESHOLD}{(default 0.01) for checking for convergence, the maximum parameter change per algorithm step}

\item{cell_types_present}{cell types (a superset of `cell_types`) to be considered as occuring often enough
to consider for gene expression contamination}

\item{fdr}{(default 0.01) false discovery rate for hypothesis testing}

\item{RCTD}{an \code{\linkS4class{RCTD}} object with annotated cell types e.g. from the \code{\link{run.RCTD}} function.}
}
\value{
an \code{\linkS4class{RCTD}} object containing the results of the RCTDE algorithm. Contains objects \code{de_results},
which contain the results of the RCTDE algorithm including `gene_fits`, which contains the results of fits on individual genes,
in addition `res_gene_list`, a list, for each cell type, of significant genes detected by RCTDE.
Additionally, the object contains `internal_vars_de`` a list of variables that are used internally by RCTDE
}
\description{
Runs RCTDE on a \code{\linkS4class{RCTD}} object (for a single predictive variable)
}
