% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{run.RCTD}
\alias{run.RCTD}
\title{Runs the RCTD pipeline on a \code{\linkS4class{RCTD}} object}
\usage{
run.RCTD(RCTD, doublet_mode = T)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object created using the \code{\link{create.RCTD}} function.}

\item{doublet_mode}{\code{logical} of whether RCTD should be run in doublet_mode.}
}
\value{
an \code{\linkS4class{RCTD}} object containing the results of the RCTD algorithm.
}
\description{
Equivalent to sequentially running the functions \code{\link{fitBulk}}, \code{\link{choose_sigma_c}}, and \code{\link{fitPixels}}
}
