% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{run.unsupervised}
\alias{run.unsupervised}
\title{Runs the unsupervised pipeline on a \code{\linkS4class{RCTD}} object}
\usage{
run.unsupervised(
  RCTD,
  doublet_mode = "doublet",
  cell_types = NULL,
  n_iter = 50,
  MIN_CHANGE = 0.001
)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object created using the \code{\link{create.RCTD}} function.}

\item{doublet_mode}{\code{character string}, either "doublet", "subtype", or "full" on which mode to run iter.optim. Please see above description.}

\item{cell_types}{the cell types used for CSIDE. If null, all cell types in \code{cell_type_info} will be chosen.}

\item{n_iter}{maximum number of optimization iterations}

\item{MIN_CHANGE}{minimum change required to terminate optimization}
}
\value{
an \code{\linkS4class{RCTD}} object containing the results of the unsupervised algorithm. Please see \code{\linkS4class{RCTD}}
documentation for more information on interpreting the content of the RCTD object.
}
\description{
Equivalent to sequentially running the functions \code{\link{choose_sigma_c}} and \code{\link{iter.optim}}
}
\details{
If in doublet mode, fits at most two cell types per pixel. It classifies each pixel as 'singlet' or 'doublet' and searches for the cell types
on the pixel. If in full mode, can fit any number of cell types on each pixel. If in subtype mode, fits only pixels belonging to given cell type(s).
}
