% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE.R
\name{runCside}
\alias{runCside}
\title{Runs cell type specific CSIDE for a general design matrix}
\usage{
runCside(
  rctd_results,
  X,
  barcodes,
  cell_types = NULL,
  gene_threshold = 5e-05,
  cell_type_threshold = 125,
  doublet_mode = TRUE,
  test_mode = "individual",
  weight_threshold = NULL,
  sigma_gene = TRUE,
  PRECISION.THRESHOLD = 0.05,
  cell_types_present = NULL,
  test_genes_sig = TRUE,
  fdr = 0.01,
  cell_type_specific = NULL,
  params_to_test = NULL,
  normalize_expr = FALSE,
  logs = FALSE,
  log_fc_thresh = 0.4,
  cell_type_filter = NULL,
  test_error = FALSE,
  initialSol = NULL
)
}
\arguments{
\item{rctd_results}{a \code{\link[SpatialExperiment]{SpatialExperiment}}
object output by the \code{\link{runRctd}} function.}

\item{X}{a matrix containing the covariates for running CSIDE. The rownames
represent pixel names and should be a subset of the pixels in the
\code{\linkS4class{SpatialRNA}} object. The columns each represent a
covariate for explaining differential expression and need to be linearly
independent.}

\item{barcodes}{the barcodes, or pixel names, of the
\code{\linkS4class{SpatialRNA}} object to be used when fitting the model.}

\item{cell_types}{the cell types used for CSIDE. If null, cell types will be
chosen with aggregate occurences of at least `cell_type_threshold`, as
aggregated by \code{\link{aggregate_cell_types}}}

\item{gene_threshold}{(default 5e-5) minimum average normalized expression
required for selecting genes}

\item{cell_type_threshold}{(default 125) min occurence of number of cells for
each cell type to be used, as aggregated by
\code{\link{aggregate_cell_types}}}

\item{doublet_mode}{(default TRUE) if TRUE, uses RCTD doublet mode weights.
Otherwise, uses RCTD full mode weights}

\item{test_mode}{(default 'individual') if 'individual', tests for DE
individually for each parameter. If 'categorical', then tests for
differences across multiple categorical parameters}

\item{weight_threshold}{(default NULL) the threshold of total normalized
weights across all cell types in \code{cell_types} per pixel to be included
in the model. Default 0.99 for doublet_mode or 0.8 for full_mode.}

\item{sigma_gene}{(default TRUE) if TRUE, fits gene specific overdispersion
parameter. If FALSE, overdispersion parameter is same across all genes.}

\item{PRECISION.THRESHOLD}{(default 0.05) for checking for convergence, the
maximum parameter change per algorithm step}

\item{cell_types_present}{cell types (a superset of `cell_types`) to be
considered as occuring often enough to consider for gene expression
contamination during the step filtering out marker genes of other cell
types.}

\item{test_genes_sig}{(default TRUE) logical controlling whether genes will
be tested for significance}

\item{fdr}{(default 0.01) false discovery rate for hypothesis testing}

\item{cell_type_specific}{(default TRUE for all covariates). A logical
vector of length the number of covariates indicating whether each
covariate's DE parameters should be cell type-specific or shared across all
cell types.}

\item{params_to_test}{(default 2 for test_mode = 'individual', all
parameters for test_mode = 'categorical'). An integer vector of parameter
indices to test. For example c(1,4,5) would test only parameters
corresponding to columns 1, 4, and 5 of the design matrix.}

\item{normalize_expr}{(default FALSE) if TRUE, constrains total gene
expression to sum to 1 in each condition. Setting normalize_expr = TRUE is
only valid for testing single parameters with test_mode = 'individual'.}

\item{logs}{(default FALSE) if TRUE, writes progress to logs/de_logs.txt}

\item{log_fc_thresh}{(default 0.4) the natural log fold change cutoff for
differential expression}

\item{test_error}{(default FALSE) if TRUE, exits after testing for error
messages without running CSIDE. If set to TRUE, this can be used to quickly
evaluate if CSIDE will run without error.}
}
\value{
an \code{\linkS4class{RctdConfig}} object containing the results of
  the CSIDE algorithm. Contains objects \code{de_results}, which contain the
  results of the CSIDE algorithm including `gene_fits`, which contains the
  results of fits on individual genes, in addition `sig_gene_list`, a list,
  for each cell type, of significant genes detected by CSIDE, whereas
  `all_gene_list` is the analogous list for all genes (including
  nonsignificant). Additionally, the object contains `internal_vars_de` a
  list of variables that are used internally by CSIDE
}
\description{
Identifies cell type specific differential expression (DE) across a general
design matrix of covariates. Uses maximum likelihood estimation to estimate
DE and standard errors for each gene and each cell type. Selects genes with
significant nonzero DE. The type of test is determined by \code{test_mode},
and the parameters tested is determined by \code{params_to_test}.
}
