% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDreplicates.R
\name{save.RCTDE.replicates}
\alias{save.RCTDE.replicates}
\title{Saves the RCTDE population-level differential expression results for a \code{\linkS4class{RCTD.replicates}} object}
\usage{
save.RCTDE.replicates(RCTD.replicates, resultsdir)
}
\arguments{
\item{RCTD.replicates}{a \code{\linkS4class{RCTD.replicates}} object containing population-level DE inference results.}

\item{resultsdir}{a directory where to save the significant gene matrices for each cell type.}
}
\description{
First, RCTDE must have been run on all replicates at the population level using e.g. the \code{\link{run.RCTDE.replicates}} function.
}
