% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_rctd_object}
\alias{simulate_rctd_object}
\title{Generate Simulated Spatial Transcriptomics Data}
\usage{
simulate_rctd_object(
  num_genes = 5000,
  num_cell_types = 3,
  num_reference_samples = 100,
  test_mixtures = pure_mixtures(num_cell_types, 25),
  prob_de = 0.1
)
}
\arguments{
\item{num_genes}{Number of genes. (Default 5000)}

\item{num_cell_types}{Number of cell types. (Default 3)}

\item{num_reference_samples}{Number of samples of each cell type to include
in the Reference object. (Default 100)}

\item{test_mixtures}{Mixtures of cell types used to create the SpatialRNA
  object, represented as a named list of named lists.

  Each inner list should have two elements:
  \itemize{
  \item{\code{n}: }{Number of samples of the mixture to create}
  \item{\code{prop}: }{Proportion of the cell types comprising the mixture,
  represented by a vector whose length is equal to the number of cell types
  and whose entries sum to 1.}
  }
  (Default is 25 samples of a pure mixture of each cell type)

  Example: \code{list(mx1=list(n=10, prop=c(0.8, 0, 0.2)))}.

  This specifies 10 samples of mixture \code{mx1}, consisting 80\% of cell
  type 1 and 20\% of cell type 3.}

\item{prob_de}{Probability that a gene will be differentially expressed.
(Default 0.1)}
}
\value{
A named list containing:
  \itemize{
  \item{\code{rctd}:}{ RCTD object constructed with a Reference object and a
  SpatialRNA object as specified above.}
  \item{\code{true_proportions}: }{Matrix where the \emph{n}th column specifies
  the true proportion of cell type \emph{n} in the SpatialRNA samples.}
  }
}
\description{
Generate Simulated Spatial Transcriptomics Data
}
\examples{
rctd_obj <- simulate_rctd_object()
}
