% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{spacexr-package}
\alias{spacexr-package}
\alias{spacexr}
\title{spacexr: an R package for assigning cell types and cell type specific differential expression to spatial transcriptomics data.}
\description{
Implementation of two computational methods in spatial transcriptomics: 1. Robust Cell Type Decomposition (RCTD). Decomposes cell type mixtures using cell type profiles, correcting for differences across sequencing technologies. 2. Cell Type-Specific Inference of Differential Expression (CSIDE). Identifies cell type-specific differential expression, accounting for localization of other cell types.
}
\section{Running RCTD}{


To get started, create a \code{\linkS4class{SpatialRNA}} object (called \code{puck} here) for the
spatial transcriptomics data and a \code{\linkS4class{Reference}} object (called \code{reference} here) for the
scRNA-seq data. Then simply run RCTD as:

\code{myRCTD <- create.RCTD(puck, reference)}

\code{myRCTD <- run.RCTD(myRCTD)}
}

\section{Running CSIDE}{


After running RCTD, create an explanatory variable (`explanatory.variable`) representing a covariate hypothesized to explain gene expression.
Then, to detect cell type-specific differential expression, simply run CSIDE as:

\code{myRCTD <- run.CSIDE.single(puck, explanatory.variable)}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrajeda/spacexr}
  \item Report bugs at \url{https://github.com/ggrajeda/spacexr/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Grajeda \email{gabriel@ds.dfci.harvard.edu}

Authors:
\itemize{
  \item Dylan Cable \email{dmcable@umich.edu}
}

}
