% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{spacexr}
\alias{spacexr}
\title{spacexr: an R package for assigning cell types and cell type specific differential expression to spatial transcriptomics data.}
\section{Running RCTD}{


To get started, create a \code{\linkS4class{SpatialRNA}} object (called \code{puck} here) for the
spatial transcriptomics data and a \code{\linkS4class{Reference}} object (called \code{reference} here) for the
scRNA-seq data. Then simply run RCTD as:

\code{myRCTD <- create.RCTD(puck, reference)}

\code{myRCTD <- run.RCTD(myRCTD)}
}

\section{Running CSIDE}{


After running RCTD, create an explanatory variable (`explanatory.variable`) representing a covariate hypothesized to explain gene expression.
Then, to detect cell type-specific differential expression, simply run CSIDE as:

\code{myRCTD <- run.CSIDE.single(puck, explanatory.variable)}
}

