% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{spatialRNA<-}
\alias{spatialRNA<-}
\title{Generic setter for spatialRNA slot}
\usage{
spatialRNA(object) <- value
}
\arguments{
\item{object}{An object with a spatialRNA slot}

\item{value}{The new value for the spatialRNA slot}
}
\value{
The updated object
}
\description{
Generic setter for spatialRNA slot
}
\keyword{internal}
