% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRNA.R
\name{SpatialRNA}
\alias{SpatialRNA}
\title{constructor of SpatialRNA object}
\usage{
SpatialRNA(
  coords,
  counts,
  nUMI = NULL,
  use_fake_coords = FALSE,
  require_int = TRUE
)
}
\arguments{
\item{coords}{A data.frame (or matrix) representing the spatial pixel locations. rownames are barcodes/pixel names,
and there should be two columns for 'x' and for 'y'.}

\item{counts}{A matrix (or dgCmatrix) representing Digital Gene Expression (DGE). Rownames should be genes
and colnames represent barcodes/pixel names.}

\item{nUMI}{Optional, a named (by pixel barcode) list of total counts or UMI's appearing at each pixel. If not provided,
nUMI will be assumed to be the total counts appearing on each pixel.}

\item{use_fake_coords}{logical, FALSE by default. If true, the 'coords' parameter will be ignored, and replaced with a placeholder
coords matrix.}

\item{require_int}{logical, TRUE by default. If true, requires counts and nUMI to be integers.

Counts should be untransformed count-level data}
}
\value{
Returns a \code{\linkS4class{SpatialRNA}} object containing the coordinates and counts
from the input files
}
\description{
constructor of SpatialRNA object
}
