% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{summarizedExperimentToReference}
\alias{summarizedExperimentToReference}
\title{Convert a SummarizedExperiment to a Reference object}
\usage{
summarizedExperimentToReference(
  reference_experiment,
  cell_type_col = "cell_type",
  require_int = TRUE,
  ref_n_cells_max = 10000,
  ref_UMI_min = 100
)
}
\arguments{
\item{reference_experiment}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object containing
annotated RNA-seq data (e.g., from snRNA-seq, scRNA-seq, or cell
type-specific bulk RNA-seq), used to learn cell type profiles. The object
must contain:
\itemize{
  \item An \code{assay} matrix of gene expression counts (genes as rows,
  cells as columns) with unique gene names as row names and unique cell
  barcodes as column names.
  \item A \code{colData} column containing cell type annotations for each
  cell (column name specified by \code{cell_type_col}).
  \item Optionally, a \code{colData} column named \code{nUMI} containing
  total UMI counts for each cell. If not provided, \code{nUMI} will be
  calculated as the column sums of the counts matrix.
}}

\item{cell_type_col}{character, name of the entry in
\code{colData(reference_experiment)} containing cell type annotations
(default: \code{"cell_type"})}

\item{require_int}{logical, whether counts and nUMI are required to be
integers (default: \code{TRUE})}

\item{ref_n_cells_max}{numeric, maximum number of cells per cell type in the
reference. Will downsample if this number is exceeded. (default: 10,000)}

\item{ref_UMI_min}{numeric, minimum UMI count for cells to be included in the
reference (default: 100)}
}
\value{
\code{\linkS4class{Reference}} object
}
\description{
Convert a SummarizedExperiment to a Reference object
}
\keyword{internal}
