% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{summarizedExperimentToSpatialRNA}
\alias{summarizedExperimentToSpatialRNA}
\title{Convert a SummarizedExperiment to a SpatialRNA object}
\usage{
summarizedExperimentToSpatialRNA(spatial_experiment, require_int = TRUE)
}
\arguments{
\item{inheritParams}{createRctd}
}
\value{
\code{\linkS4class{SpatialRNA}} object
}
\description{
Convert a SummarizedExperiment to a SpatialRNA object
}
\keyword{internal}
