% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{url_ok}
\alias{url_ok}
\title{Checks that a URL returns a 200 status code for a HEAD request}
\usage{
url_ok(url)
}
\arguments{
\item{url}{character(1)}
}
\value{
logical(1)
}
\description{
Checks that a URL returns a 200 status code for a HEAD request
}
