% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDE_plots.R
\name{write_de_summary}
\alias{write_de_summary}
\title{Saves to csv the RCTDE significant gene dataframes after running RCTDE}
\usage{
write_de_summary(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running RCTDE}

\item{datadir}{output directory}
}
\description{
Saves to csv the RCTDE significant gene dataframes after running RCTDE
}
