% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_class.R
\name{build.designmatrix.nonparam}
\alias{build.designmatrix.nonparam}
\title{Constructs a design matrix for running CSIDE nonparametrically}
\usage{
build.designmatrix.nonparam(myRCTD, barcodes = NULL, df = 15)
}
\arguments{
\item{myRCTD}{a \code{\linkS4class{CsideConfig}} object adapted from the
output of the \code{\link{runRctd}} function.}

\item{barcodes}{(default NULL) the barcodes, or pixel names, of the
\code{\linkS4class{SpatialRNA}} object to be used when creating the design
matrix.}

\item{df}{(default 15) the degrees of freedom, or number of basis functions
to be used in the design matrix}
}
\value{
A design matrix containing the covariates for running CSIDE. The
  rownames represent pixel names and are a subset of the pixels in the
  \code{\linkS4class{SpatialRNA}} object. The columns each represent a
  covariate for explaining differential expression.
}
\description{
The design matrix contains thin plate spline basis functions spanning the
space of smooth functions.
}
