% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_class.R
\name{build.designmatrix.regions}
\alias{build.designmatrix.regions}
\title{Constructs a design matrix for running CSIDE across a set of regions}
\usage{
build.designmatrix.regions(myRCTD, region_list)
}
\arguments{
\item{myRCTD}{an \code{\linkS4class{RCTD}} object with annotated cell types
e.g. from the \code{\link{run.RCTD}} function.}

\item{region_list}{a list of \code{character} vectors, where each vector
contains pixel names, or barcodes, for a single region. These pixel names
should be a subset of the pixels in the \code{\linkS4class{SpatialRNA}}
object}
}
\value{
A design matrix containing the covariates for running CSIDE. The
  rownames represent pixel names and are a subset of the pixels in the
  \code{\linkS4class{SpatialRNA}} object. The columns each represent a
  covariate for explaining differential expression.
}
\description{
The design matrix contains for each region a column of 0s and 1s representing
membership in that region.
}
