% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_utils.R
\name{cell_type_info_from_clusters}
\alias{cell_type_info_from_clusters}
\title{Generates \code{cell_type_info} from cluster assignments on a \code{\linkS4class{SpatialRNA}} object}
\usage{
cell_type_info_from_clusters(puck, clusters)
}
\arguments{
\item{puck}{a \code{\linkS4class{SpatialRNA}} object}

\item{clusters}{a \code{data.frame} with barcodes as rownames and cluster assignments under a \code{cell_types} column}
}
\value{
\code{cell_type_info}, which is ready to run the \code{\link{create.RCTD.unsupervised}} function
}
\description{
Generates \code{cell_type_info} from cluster assignments on a \code{\linkS4class{SpatialRNA}} object
}
