% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE_utils.R
\name{cell_type_info_from_de}
\alias{cell_type_info_from_de}
\title{Generates \code{cell_type_info} from CSIDE on a fitted
\code{\linkS4class{RCTD}} object}
\usage{
cell_type_info_from_de(
  RCTD,
  cell_types = NULL,
  gene_list = NULL,
  cell_type_threshold = 25
)
}
\arguments{
\item{RCTD}{a \code{\linkS4class{RCTD}} object to run CSIDE}

\item{cell_types}{the cell types used for CSIDE. If null, all cell types in
\code{cell_type_info} will be chosen.}

\item{gene_list}{the genes to compute expression for. If null, all genes will
be included.}

\item{cell_type_threshold}{(Default 10) minimum number of singlets required
per cell type.}
}
\value{
\code{cell_type_info}, which is ready to run the
  \code{\link{create.RCTD.unsupervised}} function
}
\description{
Generates \code{cell_type_info} from CSIDE on a fitted
\code{\linkS4class{RCTD}} object
}
