% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE_utils.R
\name{cell_type_info_from_singlets}
\alias{cell_type_info_from_singlets}
\title{Generates \code{cell_type_info} from singlet means on a fitted \code{\linkS4class{RCTD}} object from the original \code{\linkS4class{SpatialRNA}} object}
\usage{
cell_type_info_from_singlets(RCTD, cell_types = NULL)
}
\arguments{
\item{RCTD}{a \code{\linkS4class{RCTD}} object}

\item{cell_types}{the cell types used for CSIDE. If null, all cell types in \code{cell_type_info} will be chosen.}
}
\value{
\code{cell_type_info}, which is ready to run the \code{\link{create.RCTD.unsupervised}} function
}
\description{
Generates \code{cell_type_info} from singlet means on a fitted \code{\linkS4class{RCTD}} object from the original \code{\linkS4class{SpatialRNA}} object
}
