% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cell_types}
\alias{cell_types}
\title{Get cell type annotations}
\usage{
cell_types(x)
}
\arguments{
\item{x}{A Reference object}
}
\value{
A factor vector containing cell type annotations for each cell
}
\description{
Get cell type annotations
}
\examples{
data(rctd_simulation)

reference <- Reference(
    rctd_simulation$reference_counts,
    rctd_simulation$reference_cell_types
)

# Get cell type annotations
cell_types_vec <- cell_types(reference)

# Basic analysis of cell types
head(cell_types_vec)
levels(cell_types_vec)
table(cell_types_vec)

}
