% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{choose_sigma_c}
\alias{choose_sigma_c}
\title{Estimates sigma_c by maximum likelihood}
\usage{
choose_sigma_c(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the \code{\link{fitBulk}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} with the estimated \code{sigma_c}.
}
\description{
Estimates sigma_c by maximum likelihood
}
\examples{
data(rctd_simulation)

# Create SpatialRNA and Reference objects
spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)
reference <- Reference(
    rctd_simulation$reference_counts,
    rctd_simulation$reference_cell_types
)

# Create RCTD object
rctd <- create.RCTD(spatial_rna, reference)

rctd <- fitBulk(rctd)
rctd <- choose_sigma_c(rctd)

}
\keyword{internal}
