% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{choose_sigma_c}
\alias{choose_sigma_c}
\title{Estimates sigma_c by maximum likelihood}
\usage{
choose_sigma_c(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the
\code{\link{fitBulk}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} with the estimated
  \code{sigma_c}.
}
\description{
Estimates sigma_c by maximum likelihood
}
\examples{
data(simRctd)

# Spatial transcriptomics data
library(SpatialExperiment)
spatial_spe <- SpatialExperiment(
    assay = simRctd$spatial_rna_counts,
    spatialCoords = simRctd$spatial_rna_coords
)

# Reference data
library(SummarizedExperiment)
reference_se <- SummarizedExperiment(
    assays = list(counts = simRctd$reference_counts),
    colData = simRctd$reference_cell_types
)

# Create RCTD configuration
rctd <- createRctd(spatial_spe, reference_se, max_cores = 1)

rctd <- fitBulk(rctd)
rctd <- choose_sigma_c(rctd)

}
\keyword{internal}
