% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coords}
\alias{coords}
\alias{coords,SpatialRNA-method}
\alias{coords<-,SpatialRNA-method}
\title{Generic accessor for coords slot}
\usage{
coords(object)

\S4method{coords}{SpatialRNA}(object)

\S4method{coords}{SpatialRNA}(object) <- value
}
\arguments{
\item{object}{An object with a coords slot}
}
\value{
The coords slot of the object
}
\description{
Generic accessor for coords slot
}
\keyword{internal}
