% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{counts}
\alias{counts}
\alias{counts,SpatialRNA-method}
\alias{counts<-,SpatialRNA-method}
\alias{counts,Reference-method}
\alias{counts<-,Reference-method}
\title{Generic accessor for counts slot}
\usage{
counts(object)

\S4method{counts}{SpatialRNA}(object)

\S4method{counts}{SpatialRNA}(object) <- value

\S4method{counts}{Reference}(object)

\S4method{counts}{Reference}(object) <- value
}
\arguments{
\item{object}{An object with a counts slot}
}
\value{
The counts slot of the object
}
\description{
Generic accessor for counts slot
}
\keyword{internal}
