% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacexr.R
\name{create.RCTD}
\alias{create.RCTD}
\title{Creates an \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Reference} object and a \code{\linkS4class{SpatialRNA}} object}
\usage{
create.RCTD(
  spatialRNA,
  reference,
  max_cores = 4,
  test_mode = FALSE,
  gene_cutoff = 0.000125,
  fc_cutoff = 0.5,
  gene_cutoff_reg = 2e-04,
  fc_cutoff_reg = 0.75,
  UMI_min = 100,
  UMI_max = 2e+07,
  counts_MIN = 10,
  UMI_min_sigma = 300,
  class_df = NULL,
  CELL_MIN_INSTANCE = 25,
  cell_type_names = NULL,
  MAX_MULTI_TYPES = 4,
  keep_reference = F,
  cell_type_profiles = NULL,
  CONFIDENCE_THRESHOLD = 10,
  DOUBLET_THRESHOLD = 25
)
}
\arguments{
\item{spatialRNA}{a \code{\linkS4class{SpatialRNA}} object to run RCTD on}

\item{reference}{a \code{\linkS4class{Reference}} object scRNA-seq reference used for RCTD}

\item{max_cores}{for parallel processing, the number of cores used. If set to 1, parallel processing is not used. The system will additionally be checked for
number of available cores.}

\item{gene_cutoff}{minimum normalized gene expression for genes to be included in the platform effect normalization step.}

\item{fc_cutoff}{minimum log-fold-change (across cell types) for genes to be included in the platform effect normalization step.}

\item{gene_cutoff_reg}{minimum normalized gene expression for genes to be included in the RCTD step.}

\item{fc_cutoff_reg}{minimum log-fold-change (across cell types) for genes to be included in the RCTD step.}

\item{UMI_min}{minimum UMI per pixel included in the analysis}

\item{UMI_max}{maximum UMI per pixel included in the analysis}

\item{counts_MIN}{(default 10) minimum total counts per pixel of genes used in the analysis.}

\item{UMI_min_sigma}{minimum UMI per pixel for the \link{choose_sigma_c} function}

\item{class_df}{(optional) if not NULL, then a dataframe mapping each cell type to a cell class, so that RCTD will report confidence on the class level.}

\item{CELL_MIN_INSTANCE}{minimum number of cells required per cell type. Default 25, can be lowered if desired.}

\item{cell_type_names}{A list of cell types to be included from the reference. If NULL, uses all cell types}

\item{MAX_MULTI_TYPES}{(multi-mode only) Default 4, max number of cell types per pixel}

\item{keep_reference}{(Default FALSE) if true, keeps the \code{reference} object stored within the \code{\linkS4class{RCTD}} object}

\item{cell_type_profiles}{Default NULL, option to pass in cell type profiles in directly as a genes by cell type matrix, including gene names and cell type names.
If this option is used, reference will be ignored.}

\item{CONFIDENCE_THRESHOLD}{(Default 10) the minimum change in likelihood (compared to other cell types) necessary to determine a cell type identity with confidence}

\item{DOUBLET_THRESHOLD}{(Default 25) the penalty weight of predicting a doublet instead of a singlet for a pixel}
}
\value{
an \code{\linkS4class{RCTD}} object, which is ready to run the \code{\link{run.RCTD}} function
}
\description{
Creates an \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Reference} object and a \code{\linkS4class{SpatialRNA}} object
}
