% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{create.RCTD}
\alias{create.RCTD}
\title{Creates an  \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Seurat} object and a \code{\linkS4class{SpatialRNA}} object}
\usage{
create.RCTD(
  spatialRNA,
  reference,
  max_cores = 8,
  test_mode = FALSE,
  gene_cutoff = 0.000125,
  fc_cutoff = 0.5,
  gene_cutoff_reg = 2e-04,
  fc_cutoff_reg = 0.75,
  UMI_min = 100,
  UMI_max = 2e+05,
  class_df = NULL,
  CELL_MIN_INSTANCE = 25,
  cell_type_names = NULL
)
}
\arguments{
\item{spatialRNA}{a \code{\linkS4class{SpatialRNA}} object to run RCTD on}

\item{reference}{a \code{\linkS4class{Seurat}} object scRNA-seq reference used for RCTD}

\item{max_cores}{for parallel processing, the number of cores used. If set to 1, parallel processing is not used. The system will additionally be checked for
number of available cores.}

\item{gene_cutoff}{minimum normalized gene expression for genes to be included in the platform effect normalization step.}

\item{fc_cutoff}{minimum log-fold-change (across cell types) for genes to be included in the platform effect normalization step.}

\item{gene_cutoff_reg}{minimum normalized gene expression for genes to be included in the RCTD step.}

\item{fc_cutoff_reg}{minimum log-fold-change (across cell types) for genes to be included in the RCTD step.}

\item{UMI_min}{minimum UMI per pixel included in the analysis}

\item{UMI_max}{maximum UMI per pixel included in the analysis}

\item{class_df}{(optional) if not NULL, then a dataframe mapping each cell type to a cell class, so that RCTD will report confidence on the class level.}

\item{CELL_MIN_INSTANCE}{minimum number of cells required per cell type. Default 25, can be lowered if desired.}

\item{cell_type_names}{A list of cell types to be included from the reference. If NULL, uses all cell types}
                 
}
\value{
an \code{\linkS4class{RCTD}} object, which is ready to run the \code{\link{run.RCTD}} function
}
\description{
Creates an  \code{\linkS4class{RCTD}} object from a scRNA-seq reference \code{Seurat} object and a \code{\linkS4class{SpatialRNA}} object
}
