% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDreplicates.R
\name{create.RCTD.replicates}
\alias{create.RCTD.replicates}
\title{Creates an \code{\linkS4class{RCTD.replicates}} object across multiple \code{\linkS4class{SpatialRNA}} replicates}
\usage{
create.RCTD.replicates(
  spatialRNA.replicates,
  reference,
  replicate_names,
  group_ids = NULL,
  max_cores = 4,
  test_mode = FALSE,
  gene_cutoff = 0.000125,
  fc_cutoff = 0.5,
  gene_cutoff_reg = 2e-04,
  fc_cutoff_reg = 0.75,
  UMI_min = 100,
  UMI_max = 2e+07,
  UMI_min_sigma = 300,
  class_df = NULL,
  CELL_MIN_INSTANCE = 25,
  cell_type_names = NULL,
  MAX_MULTI_TYPES = 4,
  keep_reference = F
)
}
\arguments{
\item{spatialRNA.replicates}{a list of multiple \code{\linkS4class{SpatialRNA}} objects to run RCTD on}

\item{reference}{a \code{\linkS4class{Reference}} object scRNA-seq reference used for RCTD}

\item{replicate_names}{a \code{character} vector of names for each replicate provided in \code{spatialRNA.replicates}}

\item{group_ids}{(default constant across replicates) a named integer vector (length number of replicates) containing the group id for each replicate.
Names represent the replicate names, and replicates of the same group will be expected to be more similar than
replicates across groups}

\item{max_cores}{for parallel processing, the number of cores used. If set to 1, parallel processing is not used. The system will additionally be checked for
number of available cores.}

\item{gene_cutoff}{minimum normalized gene expression for genes to be included in the platform effect normalization step.}

\item{fc_cutoff}{minimum log-fold-change (across cell types) for genes to be included in the platform effect normalization step.}

\item{gene_cutoff_reg}{minimum normalized gene expression for genes to be included in the RCTD step.}

\item{fc_cutoff_reg}{minimum log-fold-change (across cell types) for genes to be included in the RCTD step.}

\item{UMI_min}{minimum UMI per pixel included in the analysis}

\item{UMI_max}{maximum UMI per pixel included in the analysis}

\item{UMI_min_sigma}{minimum UMI per pixel for the \link{choose_sigma_c} function}

\item{class_df}{(optional) if not NULL, then a dataframe mapping each cell type to a cell class, so that RCTD will report confidence on the class level.}

\item{CELL_MIN_INSTANCE}{minimum number of cells required per cell type. Default 25, can be lowered if desired.}

\item{cell_type_names}{A list of cell types to be included from the reference. If NULL, uses all cell types}

\item{MAX_MULTI_TYPES}{(multi-mode only) Default 4, max number of cell types per pixel}

\item{keep_reference}{(Default FALSE) if true, keeps the \code{reference} object stored within the \code{\linkS4class{RCTD}} object}

\item{cell_type_info}{Default NULL, option to pass in \code{cell_type_info} directly}
}
\value{
an \code{\linkS4class{RCTD.replicates}} object, which is ready to run the \code{\link{run.RCTD.replicates}} function
}
\description{
Applies the \code{\link{create.RCTD}} function for each \code{\linkS4class{SpatialRNA}} replicate inputted using a
scRNA-seq reference \code{Reference} object.
}
