% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{fitBulk}
\alias{fitBulk}
\title{Performs Platform Effect Normalization:}
\usage{
fitBulk(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the \code{\link{create.RCTD}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} object normalized for platform effects.
}
\description{
Estimates bulk cell type composition and uses this
to estimate platform effects and normalize cell type proportions
}
