% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{fitBulk}
\alias{fitBulk}
\title{Performs Platform Effect Normalization:}
\usage{
fitBulk(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the \code{\link{create.RCTD}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} object normalized for platform effects.
}
\description{
Estimates bulk cell type composition and uses this
to estimate platform effects and normalize cell type proportions
}
\examples{
data(rctd_simulation)

# Spatial transcriptomics data
library(SpatialExperiment)
spatial_spe <- SpatialExperiment(
    assay = rctd_simulation$spatial_rna_counts,
    spatialCoords = rctd_simulation$spatial_rna_coords
)

# Reference data
library(SummarizedExperiment)
reference_se <- SummarizedExperiment(
    assays = list(counts = rctd_simulation$reference_counts),
    colData = rctd_simulation$reference_cell_types
)

# Create RCTD configuration
rctd <- create.RCTD(spatial_spe, reference_se, max_cores = 1)
rctd <- fitBulk(rctd)

}
\keyword{internal}
