% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{fitBulk}
\alias{fitBulk}
\title{Performs Platform Effect Normalization:}
\usage{
fitBulk(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the
\code{\link{createRctd}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} object normalized for platform
  effects.
}
\description{
Estimates bulk cell type composition and uses this to estimate platform
effects and normalize cell type proportions
}
\examples{
data(simRctd)

# Spatial transcriptomics data
library(SpatialExperiment)
spatial_spe <- SpatialExperiment(
    assay = simRctd$spatial_rna_counts,
    spatialCoords = simRctd$spatial_rna_coords
)

# Reference data
library(SummarizedExperiment)
reference_se <- SummarizedExperiment(
    assays = list(counts = simRctd$reference_counts),
    colData = simRctd$reference_cell_types
)

# Create RCTD configuration
rctd <- createRctd(spatial_spe, reference_se, max_cores = 1)
rctd <- fitBulk(rctd)

}
\keyword{internal}
