% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_effect_normalization.R
\name{fitBulk}
\alias{fitBulk}
\title{Performs Platform Effect Normalization:}
\usage{
fitBulk(RCTD)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object after running the \code{\link{create.RCTD}} function.}
}
\value{
Returns an \code{\linkS4class{RCTD}} object normalized for platform effects.
}
\description{
Estimates bulk cell type composition and uses this
to estimate platform effects and normalize cell type proportions
}
\examples{
data(rctd_simulation)

# Create SpatialRNA and Reference objects
spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)
reference <- Reference(
    rctd_simulation$reference_counts,
    rctd_simulation$reference_cell_types
)

# Create RCTD object
rctd <- create.RCTD(spatial_rna, reference, max_cores = 1)
rctd <- fitBulk(rctd)

}
\keyword{internal}
