% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{gather_results}
\alias{gather_results}
\title{Collects RCTD results into a dataframe}
\usage{
gather_results(RCTD, results)
}
\arguments{
\item{RCTD}{RCTD object}

\item{results}{List of RCTD results for each pixel, output by
\code{\link{process_beads_batch}}}
}
\value{
RCTD results organized into a dataframe
}
\description{
Collects RCTD results into a dataframe
}
\keyword{internal}
