% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE_init.R
\name{initialize.clusters}
\alias{initialize.clusters}
\title{Generates \code{cell_type_info} from Seurat clustering on a
\code{\linkS4class{SpatialRNA}} object}
\usage{
initialize.clusters(puck, resolution = 0.7, SCT = TRUE)
}
\arguments{
\item{puck}{a \code{\linkS4class{SpatialRNA}} object to run clustering on}

\item{resolution}{(Default 0.7) the resolution to be used during Seurat
clustering}

\item{SCT}{whether or not to use \code{SCTransform} for count normalization}
}
\value{
\code{cell_type_info}, which is ready to run the
  \code{\link{create.RCTD.unsupervised}} function
}
\description{
Generates \code{cell_type_info} from Seurat clustering on a
\code{\linkS4class{SpatialRNA}} object
}
