% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_init.R
\name{initialize.subtypes}
\alias{initialize.subtypes}
\title{Creates an \code{\linkS4class{RCTD}} object ready to be run on subtype mode from a fitted \code{\linkS4class{RCTD}} object}
\usage{
initialize.subtypes(
  RCTD,
  cell_types,
  resolution = 0.7,
  SCT = T,
  gene_list_tot = NULL
)
}
\arguments{
\item{RCTD}{a \code{\linkS4class{RCTD}} object with cell types fitted}

\item{cell_types}{cell supertypes to find subtypes for}

\item{resolution}{(Default 0.7) the resolution to be used during Seurat clustering for subtype initialization}

\item{SCT}{whether or not to use \code{SCTransform} for count normalization}

\item{gene_list_tot}{gene list to be used for subtype mode. If null, uses highly expressed supertype genes and differentially expressed subtype cluster genes}
}
\value{
an \code{\linkS4class{RCTD}} object, which is ready to run the \code{\link{run.unsupervised}} function on subtype mode
}
\description{
Creates an \code{\linkS4class{RCTD}} object ready to be run on subtype mode from a fitted \code{\linkS4class{RCTD}} object
}
