% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE_init.R
\name{initialize.subtypes}
\alias{initialize.subtypes}
\title{Creates an \code{\linkS4class{RctdConfig}} object ready to be run on subtype
mode from a fitted \code{\linkS4class{RctdConfig}} object}
\usage{
initialize.subtypes(RCTD, cell_types, resolution = 0.7, gene_list = NULL)
}
\arguments{
\item{RCTD}{a \code{\linkS4class{RctdConfig}} object with cell types fitted}

\item{cell_types}{cell supertypes to find subtypes for}

\item{resolution}{(Default 0.7) the resolution to be used during Seurat
clustering for subtype initialization}

\item{gene_list}{gene list to be used for subtype mode. If null, uses highly
expressed supertype genes}
}
\value{
an \code{\linkS4class{RctdConfig}} object, which is ready to run the
  \code{\link{run.unsupervised}} function on subtype mode
}
\description{
Creates an \code{\linkS4class{RctdConfig}} object ready to be run on subtype
mode from a fitted \code{\linkS4class{RctdConfig}} object
}
