% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_SQ_all}
\alias{load_SQ_all}
\title{Retrieves SQ matrices from cache, populating the cache if necessary.}
\usage{
load_SQ_all()
}
\value{
list of matrices
}
\description{
Retrieves SQ matrices from cache, populating the cache if necessary.
}
\keyword{internal}
