% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDE_plots.R
\name{make_all_de_plots}
\alias{make_all_de_plots}
\title{Makes all RCTDE plots on RCTD object, after running RCTDE}
\usage{
make_all_de_plots(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running RCTDE}

\item{datadir}{output directory}
}
\description{
Makes all RCTDE plots on RCTD object, after running RCTDE
}
