% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{make_all_de_plots}
\alias{make_all_de_plots}
\title{Makes all CSIDE plots on RCTD object, after running CSIDE}
\usage{
make_all_de_plots(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running CSIDE}

\item{datadir}{output directory}
}
\description{
Makes all CSIDE plots on RCTD object, after running CSIDE
}
