% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_loading.R
\name{make_cache}
\alias{make_cache}
\title{Returns a stateful function that stores the most recent non-null argument and
returns it for NULL values.}
\usage{
make_cache()
}
\value{
cache function
}
\description{
This is effectively used to manage global variables.
}
\keyword{internal}
