% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{make_de_plots_quant}
\alias{make_de_plots_quant}
\title{Makes quantitative CSIDE plots on RCTD object, after running CSIDE}
\usage{
make_de_plots_quant(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running CSIDE}

\item{datadir}{output directory}
}
\description{
Makes quantitative CSIDE plots on RCTD object, after running CSIDE
}
