% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{make_de_plots_regions}
\alias{make_de_plots_regions}
\title{Makes spatial gene CSIDE plots (colored by discrete regions) on RCTD object, after running CSIDE}
\usage{
make_de_plots_regions(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running CSIDE}

\item{datadir}{output directory}
}
\description{
Makes spatial gene CSIDE plots (colored by discrete regions) on RCTD object, after running CSIDE
}
