% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDE_plots.R
\name{make_de_plots_spatial}
\alias{make_de_plots_spatial}
\title{Makes spatial gene RCTDE plots (colored by two discrete regions) on RCTD object, after running RCTDE}
\usage{
make_de_plots_spatial(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running RCTDE}

\item{datadir}{output directory}
}
\description{
Makes spatial gene RCTDE plots (colored by two discrete regions) on RCTD object, after running RCTDE
}
