% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{nUMI}
\alias{nUMI}
\alias{nUMI,SpatialRNA-method}
\alias{nUMI<-,SpatialRNA-method}
\alias{nUMI,Reference-method}
\alias{nUMI<-,Reference-method}
\title{Generic accessor for nUMI slot}
\usage{
nUMI(object)

\S4method{nUMI}{SpatialRNA}(object)

\S4method{nUMI}{SpatialRNA}(object) <- value

\S4method{nUMI}{Reference}(object)

\S4method{nUMI}{Reference}(object) <- value
}
\arguments{
\item{object}{An object with a nUMI slot}
}
\value{
The nUMI slot of the object
}
\description{
Generic accessor for nUMI slot
}
\keyword{internal}
