% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_type_weight}
\alias{plot_cell_type_weight}
\title{Plot pixel proportions for a specific cell type}
\usage{
plot_cell_type_weight(
  rctd_se,
  cell_type,
  weights_assay = "weights",
  size = 10,
  stroke = 1,
  alpha = 1,
  low = "white",
  high = "red",
  plot_title = NA
)
}
\arguments{
\item{rctd_se}{SummarizedExperiment containing RCTD results}

\item{cell_type}{character(1) name of cell type to plot}

\item{weights_assay}{character(1) name of the weights assay (default:
"weights")}

\item{size}{numeric(1) size of the points (default: 10)}

\item{stroke}{numeric(1) border width of the points (default: 1)}

\item{alpha}{numeric(1) point transparency between 0 and 1 (default: 1)}

\item{low}{character(1) color for the low end of the proportion color scale
(default: "white")}

\item{high}{character(1) color for the high end of the proportion color scale
(default: "red")}

\item{plot_title}{character(1) plot title. (default: NA)}
}
\value{
ggplot object showing the proportion of a specified cell type at each
  pixel
}
\description{
Creates a visualization showing how the proportion of a specific cell type
varies across space, represented by point color intensity.
}
