% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cond_occur}
\alias{plot_cond_occur}
\title{Barplot of the confident counts for each cell type}
\usage{
plot_cond_occur(cell_type_names, resultsdir, weights, puck)
}
\arguments{
\item{cell_type_names}{list of cell type names}

\item{resultsdir}{output directory}

\item{weights}{a dataframe of RCTD output weights (see \code{\link{gather_results}})}

\item{puck}{an object of type \linkS4class{SpatialRNA}}
}
\value{
returns \code{\link{ggplot2}} object
}
\description{
Plots the number of confident labels in 'full_mode'. Saves as 'cell_type_occur.pdf'
}
