% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_doub_occur_stack}
\alias{plot_doub_occur_stack}
\title{Plots doublet co-occurances}
\usage{
plot_doub_occur_stack(doub_occur, resultsdir, cell_type_names)
}
\arguments{
\item{doub_occur}{a table of occurances of doublets}

\item{resultsdir}{output directory}

\item{cell_type_names}{list of cell type names}
}
\value{
returns \code{\link{ggplot2}} object
}
\description{
Plots the doublet co-occurances. Saves as 'doublet_stacked_bar.pdf'
}
