% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{plot_gene_two_regions}
\alias{plot_gene_two_regions}
\title{Makes a spatial plot of gene expression for a particular gene
This plot is colored by two discrete regions based on high or low explanatory variable values.
Bold points represent expressed, whereas unbold points represent pixels not expressing the gene.}
\usage{
plot_gene_two_regions(
  myRCTD,
  gene,
  cell_type,
  min_UMI = 200,
  expr.thresh = 0,
  exvar_thresh = 0.5
)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running CSIDE}

\item{gene}{gene to be plotted}

\item{cell_type}{cell_type to be plotted (only single cell type pixels)}

\item{min_UMI}{(default 200) minimum UMI for pixels that are included}

\item{expr.thresh}{(default 0) the minimum expression threshold to clear to be considered to be expressed}

\item{exvar_thresh}{threshold of the explanatory variable in order for points to be sorted into the two regions}
}
\value{
gene expression plot
}
\description{
Makes a spatial plot of gene expression for a particular gene
This plot is colored by two discrete regions based on high or low explanatory variable values.
Bold points represent expressed, whereas unbold points represent pixels not expressing the gene.
}
