% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRCTD.R
\name{process_beads_batch}
\alias{process_beads_batch}
\title{Runs RCTD in doublet mode on \code{puck}}
\usage{
process_beads_batch(
  cell_type_info,
  gene_list,
  puck,
  class_df = NULL,
  constrain = T,
  MAX_CORES = 8,
  MIN.CHANGE = 0.001
)
}
\arguments{
\item{cell_type_info}{cell type information and profiles of each cell, calculated from the scRNA-seq
reference (see \code{\link{get_cell_type_info}})}

\item{gene_list}{a list of genes to be used for RCTD}

\item{puck}{an object of type \linkS4class{SpatialRNA}, the target dataset}

\item{class_df}{A dataframe returned by \code{\link{get_class_df}} to map cell types to
classes}

\item{constrain}{logical whether to constrain the weights to sum to one on each pixel}

\item{max_cores}{number of cores to use (will use parallel processing if more than one).}
}
\value{
Returns \code{results}, a list of RCTD results for each pixel, which can be organized by
feeding into \code{\link{gather_results}}
}
\description{
Then, computes cell type proportions for each pixel in \code{puck}.
Classifies each pixel as 'singlet' or 'doublet' and searches for the cell types
on the pixel
}
