% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RCTD-class}
\alias{RCTD-class}
\title{An S4 class used to run the RCTD algorithm}
\description{
Created using the \code{\link{create.RCTD}} function, a user can run RCTD using the \code{\link{run.RCTD}} function.
}
\section{Slots}{

\describe{
\item{\code{spatialRNA}}{a \code{\linkS4class{SpatialRNA}} object containing the Spatial RNA dataset to be processed}

\item{\code{reference}}{a \code{\linkS4class{Reference}} object containing the cell type-labeled single cell reference}

\item{\code{config}}{a list of configuration options, set using the \code{\link{create.RCTD}} function}

\item{\code{cell_type_info}}{a named list of cell type profiles (means), containing two elements: \code{info}, directly calculated from the scRNA-seq reference, and
\code{renorm}, renormalized the match the SpatialRNA dataset.}

\item{\code{internal_vars}}{a list of internal variables used by RCTD's computation}

\item{\code{results}}{(created after running RCTD) a list of results_df (a dataframe of RCTD results in doublet mode),
weights (a dataframe of RCTD predicted weights in full mode), and weights_doublet (a
dataframe of predicted weights in doublet mode, with cell type information in results_df)}
}}

