% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Reference-class}
\alias{Reference-class}
\title{An S4 class to represent Single-Cell RNA-seq reference}
\description{
An S4 class to represent Single-Cell RNA-seq reference
}
\section{Slots}{

\describe{
\item{\code{cell_types}}{a factor of cell type identities for each cell}

\item{\code{counts}}{a sparse matrix of raw counts for each gene (rowname)
and each cell (colnames or barcodes)}

\item{\code{nUMI}}{an atomic vector of numeric UMI counts per cell}
}}

