% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference}
\alias{Reference}
\title{constructor of \code{\linkS4class{Reference}} object}
\usage{
Reference(
  counts,
  cell_types,
  nUMI = NULL,
  require_int = TRUE,
  n_max_cells = 10000,
  min_UMI = 100
)
}
\arguments{
\item{counts}{A matrix (or dgCmatrix) representing Digital Gene Expression (DGE). Rownames should be genes
and colnames represent barcodes/cell names.}

\item{cell_types}{A named (by cell barcode) factor of cell type for each cell. The 'levels' of the factor would be the possible
cell type identities.}

\item{nUMI}{Optional, a named (by cell barcode) list of total counts or UMI's appearing at each pixel. If not provided,
nUMI will be assumed to be the total counts appearing on each pixel.}

\item{n_max_cells}{(default 10,000) the maximum number of cells per cell type. Will downsample if this number is exceeded.

Counts should be untransformed count-level data}

\item{min_UMI}{(default 100) minimum UMI count for cells to be included in the reference.}
}
\value{
Returns a \code{\linkS4class{Reference}} object containing the counts matrix, cell type labels, and UMI vector
from the input files
}
\description{
constructor of \code{\linkS4class{Reference}} object
}
