% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference}
\alias{Reference}
\title{constructor of \code{\linkS4class{Reference}} object}
\usage{
Reference(counts, cell_types, nUMI = NULL)
}
\arguments{
\item{counts}{A matrix (or dgCmatrix) representing Digital Gene Expression (DGE). Rownames should be genes
and colnames represent barcodes/cell names.}

\item{cell_types}{A named (by cell barcode) factor of cell type for each cell. The 'levels' of the factor would be the possible
cell type identities.}

\item{nUMI}{Optional, a named (by cell barcode) list of total counts or UMI's appearing at each pixel. If not provided,
nUMI will be assumed to be the total counts appearing on each pixel.

Counts should be untransformed count-level data}
}
\value{
Returns a \code{\linkS4class{Reference}} object containing the counts matrix, cell type labels, and UMI vector
from the input files
}
\description{
constructor of \code{\linkS4class{Reference}} object
}
