% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTDreplicates.R
\name{run.RCTD.replicates}
\alias{run.RCTD.replicates}
\title{Runs the RCTD pipeline on a \code{\linkS4class{RCTD.replicates}} object}
\usage{
run.RCTD.replicates(RCTD.replicates, doublet_mode = "doublet")
}
\arguments{
\item{RCTD.replicates}{an \code{\linkS4class{RCTD.replicates}} object created using the \code{\link{create.RCTD.replicates}} function.}

\item{doublet_mode}{\code{character string}, either "doublet", "multi", or "full" on which mode to run RCTD. Please see above description.}
}
\value{
an \code{\linkS4class{RCTD.replicates}} object containing the results of the RCTD algorithm. Please see \code{\linkS4class{RCTD.replicates}}
and \code{\linkS4class{RCTD}} documentation for more information on interpreting the content of this object.
}
\description{
For each \code{\linkS4class{SpatialRNA}} replicate in the \code{RCTD.replicates} object, runs the \code{\link{run.RCTD}}
function to assign cell types.
}
\details{
If in doublet mode, fits at most two cell types per pixel. It classifies each pixel as 'singlet' or 'doublet' and searches for the cell types
on the pixel. If in full mode, can fit any number of cell types on each pixel. In multi mode, cell types are added using a greedy algorithm,
up to a fixed number.
}
