% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICE.R
\name{run.RICE}
\alias{run.RICE}
\title{Runs the unsupervised pipeline on a \code{\linkS4class{RCTD}} object}
\usage{
run.RICE(
  RCTD,
  doublet_mode = "doublet",
  cell_types = NULL,
  n_iter = 50,
  MIN_CHANGE = 0.001,
  return_list = F
)
}
\arguments{
\item{RCTD}{an \code{\linkS4class{RCTD}} object created using the \code{\link{create.RCTD}} function.}

\item{doublet_mode}{\code{character string}, either "doublet", "subtype", or "full" on which mode to run iterOptim. Please see above description.}

\item{n_iter}{maximum number of optimization iterations}

\item{MIN_CHANGE}{minimum change required to terminate optimization}
}
\value{
an \code{\linkS4class{RCTD}} object containing the results of the unsupervised algorithm. Please see \code{\linkS4class{RCTD}}
documentation for more information on interpreting the content of the RCTD object.
}
\description{
Equivalent to sequentially running the functions \code{\link{choose_sigma_c}} and \code{\link{iterOptim}}
}
\details{
If in doublet mode, fits at most two cell types per pixel. It classifies each pixel as 'singlet' or 'doublet' and searches for the cell types
on the pixel. If in full mode, can fit any number of cell types on each pixel.
}
