% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_utils.R
\name{set_cell_types_assigned}
\alias{set_cell_types_assigned}
\title{If cell types have been assigned to the RCTD object, running this function
will toggle the cell_types_assigned variable, which enables CSIDE to be run.}
\usage{
set_cell_types_assigned(myRCTD)
}
\arguments{
\item{myRCTD}{an \code{\linkS4class{RctdConfig}} object adapted from the
output of the \code{\link{runRctd}} function.}
}
\value{
the `myRCTD` object with cell_types_assigned set to TRUE
}
\description{
If cell types have been assigned to the RCTD object, running this function
will toggle the cell_types_assigned variable, which enables CSIDE to be run.
}
