% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{spacexr-package}
\alias{spacexr-package}
\alias{spacexr}
\title{spacexr: Cell Type Identification in Spatial Transcriptomics}
\description{
Robust Cell Type Decomposition (RCTD) is a computational method for deconvolving cell type mixtures in spatial transcriptomics data. RCTD learns cell type profiles from annotated RNA sequencing (RNA-seq) reference data and uses these profiles to identify cell types in spatial transcriptomic pixels while accounting for platform-specific effects. The package provides three analysis modes: doublet detection (up to 2 cell types per pixel), multi mode with greedy cell type assignment (up to a specified number of cell types per pixel), and full decomposition without restrictions. This is the Bioconductor version of the original spacexr package, which can be found here: https://github.com/dmcable/spacexr. Future releases will include an implementation of CSIDE.
}
\section{Running RCTD}{


  To get started, create a \code{\link[SpatialExperiment]{SpatialExperiment}}
  object (called \code{spatial} here) for the spatial transcriptomics data
  and a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
  (called \code{reference} here) for the RNA-seq data. Then simply run RCTD
  as:

  \code{rctd_config <- createRctd(spatial, reference)}

  \code{results <- runRctd(rctd_config)}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrajeda/spacexr}
  \item Report bugs at \url{https://github.com/ggrajeda/spacexr/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Grajeda \email{gabriel.grajeda@gmail.com} (\href{https://orcid.org/0009-0003-7242-7476}{ORCID})

Authors:
\itemize{
  \item Dylan Cable \email{dmcable@umich.edu}
}

}
