% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SpatialRNA-class}
\alias{SpatialRNA-class}
\title{An S4 class to represent Spatial Transcriptomic data}
\description{
An S4 class to represent Spatial Transcriptomic data
}
\section{Slots}{

\describe{
\item{\code{coords}}{a dataframe with x and y coordinates of each pixel}

\item{\code{counts}}{a sparse matrix of raw counts for each gene (rowname)
and each pixel (colnames or barcodes)}

\item{\code{n_cell_type}}{the number of cell types}

\item{\code{cell_type_names}}{a list of cell type names}

\item{\code{nUMI}}{a named list (by barcode) of total UMIs per pixel}

\item{\code{cell_labels}}{a factor of cell type labels for each pixel}
}}

