% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SpatialRNA-class}
\alias{SpatialRNA-class}
\alias{coords}
\alias{counts}
\alias{nUMI}
\alias{show,SpatialRNA-method}
\title{Spatial transcriptomics data}
\usage{
coords(object)

counts(object)

nUMI(object)

\S4method{show}{SpatialRNA}(object)
}
\description{
A class representing spatial transcriptomics data, where gene expression is
measured at fixed locations called "pixels" (also known as "spots" or
"beads"), which may contain mixtures of multiple cells. RCTD estimates the
proportions of different cell types in each pixel.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data frame (or matrix) containing x and y coordinates for each
pixel (identified by barcode)}

\item{\code{counts}}{sparse matrix of gene expression counts, with genes as rows and
pixels as columns (named by pixel barcode)}

\item{\code{nUMI}}{numeric vector of total UMI counts per pixel (identified by
barcode)}
}}

\examples{
data(rctd_simulation)

# Create SpatialRNA object
spatial_rna <- SpatialRNA(
    rctd_simulation$spatial_rna_coords,
    rctd_simulation$spatial_rna_counts
)

# Access object slots using generics
head(coords(spatial_rna))  # Get spatial coordinates
dim(counts(spatial_rna))   # Get gene expression matrix dimensions
head(nUMI(spatial_rna))    # Get total UMI counts per pixel

}
