% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{spatialRNA}
\alias{spatialRNA}
\alias{spatialRNA,RctdConfig-method}
\alias{spatialRNA<-,RctdConfig-method}
\title{Generic accessor for spatialRNA slot}
\usage{
spatialRNA(object)

\S4method{spatialRNA}{RctdConfig}(object)

\S4method{spatialRNA}{RctdConfig}(object) <- value
}
\arguments{
\item{object}{An object with a spatialRNA slot}
}
\value{
The spatialRNA slot of the object
}
\description{
Generic accessor for spatialRNA slot
}
\keyword{internal}
