% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_utils.R
\name{weights_change}
\alias{weights_change}
\title{Calculates the weight change between two \code{\linkS4class{RCTD}} objects}
\usage{
weights_change(RCTD1, RCTD2)
}
\arguments{
\item{RCTD1}{a \code{\linkS4class{RCTD}} object}

\item{RCTD2}{a \code{\linkS4class{RCTD}} object}
}
\value{
\code{cell_type_info}, which is ready to run the \code{\link{create.RCTD.unsupervised}} function
}
\description{
Used during \code{\link{iter.optim}} as the termination criterion
}
