% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_plots.R
\name{write_de_summary}
\alias{write_de_summary}
\title{Saves to csv the CSIDE significant gene dataframes after running CSIDE}
\usage{
write_de_summary(myRCTD, datadir)
}
\arguments{
\item{myRCTD}{\code{\linkS4class{RCTD}} object containing \code{de_results}, after running CSIDE}

\item{datadir}{output directory}
}
\description{
Saves to csv the CSIDE significant gene dataframes after running CSIDE
}
